% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{addComponent}
\alias{addComponent}
\title{Creates a \code{pipe} object}
\usage{
addComponent(newComponent, pipeline)
}
\arguments{
\item{newComponent}{\code{pipeline} or \code{module} object to be added}

\item{pipeline}{\code{pipeline} to be amended}

\item{startComponent}{Name of start component}

\item{startOutput}{Name of start output}

\item{endComponent}{Name of end module}

\item{endInput}{Name of end input
Add a new component to a pipeline

This function adds a new \code{module} or \code{pipeline} to a
\code{pipeline}'s components.}
}
\value{
\code{pipeline} object
}
\description{
Constructor method for a \code{pipe} object which connects the
\code{startComponent}'s \code{startOutput} to the
\code{endComponent}'s \code{endInput}.
}
\details{
Accessor methods are defined to extract \code{start} and \code{end}
elements.
}
\examples{
## create a pipeline with one module
mod1 <- module(name = "setX", language = "R",
               description = "sets the value of x",
               outputs = list(
                   moduleOutput(
                       name = "x",
                       vessel = internalVessel("x"),
                       format = ioFormat("R character string"))),
               sources = list(
                   moduleSource(
                       vessel = scriptVessel("x <- \\"set\\""))))
pline1 <- pipeline(name = "trivialpipeline", components = list(mod1))
## create a new module
mod2 <- module("showY", language = "R",
               description = "displays the value of Y",
               inputs = list(
                   moduleInput(
                       name = "y",
                       vessel = internalVessel("y"),
                       format = ioFormat("R character string"))),
               sources = list(
                   moduleSource(
                       vessel = scriptVessel("print(y)"))))
## add new module to pipeline
pline1 <- addComponent(mod2, pline1)
}
\seealso{
\code{pipeline}, \code{module}
}

