% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addPipe}
\alias{addPipe}
\title{add a pipe to a pipeline}
\usage{
addPipe(newPipe, pipeline)
}
\arguments{
\item{newPipe}{\code{pipe} object}

\item{pipeline}{\code{pipeline} object}
}
\value{
\code{pipeline} object
}
\description{
Add a \code{pipe} object to a \code{pipeline}
}

