% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{amendSearchPaths}
\alias{amendSearchPaths}
\title{Amend search paths}
\usage{
amendSearchPaths(newPaths, pathsToAmend = defaultSearchPaths)
}
\arguments{
\item{newPaths}{Character string of paths to be added}

\item{pathsToAmend}{Character string of paths to be amended}
}
\value{
Character string of amended search paths
}
\description{
Append, prepend, or replace search paths with a new set.
}
\details{
If \code{newPaths} ends with the | character, new search paths will be
prepended to \code{pathsToAmend}.

If \code{newPaths} begins with | character, new search paths will be
appended to \code{pathsToAmend}.

Otherwise \code{newPaths} replaces \code{pathsToAmend}.
}

