% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{as_length1_char}
\alias{as_length1_char}
\title{Coerce a vector to a character vector of length 1}
\usage{
as_length1_char(x)
}
\arguments{
\item{x}{vector to be coerced}
}
\value{
character vector of length one
}
\description{
Coerce a vector to a character vector of length 1
}

