% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{buildModuleHost}
\alias{buildModuleHost}
\title{Build a host string from a parsed host list}
\usage{
buildModuleHost(parsedHost)
}
\arguments{
\item{parsedHost}{list generated by \code{parseModuleHost}}
}
\value{
character string describing module host
}
\description{
Build a host string from a parsed host list
}

