% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{calculateModuleOutputs}
\alias{calculateModuleOutputs}
\title{Calculate \code{output} objects produced by a \code{module}}
\usage{
calculateModuleOutputs(module, outputDirectory)
}
\arguments{
\item{module}{\code{module} object}

\item{outputDirectory}{file location for outputs}
}
\value{
list of \code{output} objects
}
\description{
Calculate \code{output} objects produced by a \code{module}
}

