% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{calculateOutputs.module}
\alias{calculateOutputs.module}
\title{calculate output objects produced by a module}
\usage{
\method{calculateOutputs}{module}(componentValue, outputDirectory)
}
\arguments{
\item{componentValue}{\code{module} object from \code{component}
\code{value} slot}

\item{outputDirectory}{file location for pipeline execution output}
}
\description{
calculate output objects produced by a module
}

