% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{checkOutputObject}
\alias{checkOutputObject}
\title{Checks a module output object has been created.}
\usage{
checkOutputObject(output, internalExtension)
}
\arguments{
\item{output}{\code{moduleOutput} object}

\item{internalExtension}{file extension for serialized internal language
object}
}
\value{
named list containing:
\itemize{
  \item name: object name
  \item type: object vessel type
  \item object: output object
}
}
\description{
Checks a module output object has been created.
}
\details{
Will produce an error if the object does not exist.
}

