% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/component.R, R/pipeline.R
\name{getName.component}
\alias{component}
\alias{getName.component}
\title{Create a component object}
\usage{
\method{getName}{component}(x)

component(name, type = NULL, value = NULL)

component(name, type = NULL, value = NULL)
}
\arguments{
\item{name}{Name of component}

\item{type}{Character value; \dQuote{pipeline} or \dQuote{module}}

\item{value}{\code{vessel}, \code{pipeline} or \code{module} object}

\item{name}{Name of component}

\item{type}{Character value; \dQuote{pipeline} or \dQuote{module}}

\item{value}{\code{vessel}, \code{pipeline} or \code{module} object}
}
\value{
\code{component} list containing:
\describe{
  \item{name}{component name}
  \item{type}{component type}
  \item{value}{\code{vessel}, \code{pipeline} or \code{module} object}
}

\code{component} list containing:
\describe{
  \item{name}{component name}
  \item{type}{component type}
  \item{value}{\code{vessel}, \code{pipeline} or \code{module} object}
}
}
\description{
Create a \code{component} object for use in a \code{pipeline}.

Create a \code{component} object for use in a \code{pipeline}.
}
\details{
A component contains either a \code{pipeline} or
    \code{module} object, or a \code{vessel} object which
    references and object which can be loaded as one of these.

The object passed to \code{value} can be any type of vessel except
\code{internalVessel} or \code{scriptVessel}.

A component contains either a \code{pipeline} or
    \code{module} object, or a \code{vessel} object which
    references and object which can be loaded as one of these.

The object passed to \code{value} can be any type of vessel except
\code{internalVessel} or \code{scriptVessel}.
}
\seealso{
\code{pipeline}, \code{module}

\code{pipeline}, \code{module}
}

