% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{componentName}
\alias{componentName}
\title{return the name of a component}
\usage{
componentName(component)
}
\arguments{
\item{component}{\code{module} or \code{pipeline} object}
}
\value{
character value
}
\description{
Returns the name of a \code{module} or \code{pipeline}
}

