% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/component.R
\name{componentPath}
\alias{componentPath}
\title{Calculate a component's output path}
\usage{
componentPath(component, pipelinePath)
}
\arguments{
\item{component}{\code{component} object}

\item{pipelinePath}{output path of parent pipeline}
}
\value{
output path as character
}
\description{
Calculate a component's output path
}

