% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{componentResult}
\alias{componentResult}
\title{Result of \code{runModule} or \code{runPipeline}}
\description{
Result of \code{runModule} or \code{runPipeline}
}
\seealso{
\code{moduleResult}, \code{pipelineResult}
}

