% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{componentToXML}
\alias{componentToXML}
\title{Convert a component to XML}
\usage{
componentToXML(component, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{component}{\code{component} object}

\item{namespaceDefinitions}{As named character vector}

\item{parent}{XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{xmlNode} object
}
\description{
Convert a \code{component} object into the corresponding openapi XML
format.
}
\seealso{
\code{XML::newXMLNode}
}

