% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{createHostDirectory}
\alias{createHostDirectory}
\title{Create module output directory on host}
\usage{
createHostDirectory(host)
}
\arguments{
\item{host}{remote host list}
}
\value{
0 if success
}
\description{
Create module output directory on host
}

