% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{defaultIdfile}
\alias{defaultIdfile}
\title{Default idfile for ssh to remote hosts.}
\format{\preformatted{ chr "/home/ashley/uoastorage/openapi/conduit/inst/conduit.key"
}}
\usage{
defaultIdfile
}
\description{
Availabe at \code{system.file("conduit.key", package = "conduit")}
}
\details{
Corresponding public key at
\code{system.file("conduit.key.pub", package = "conduit")}
}
\keyword{datasets}

