% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{defaultSearchPaths}
\alias{defaultSearchPaths}
\title{The default Search Paths}
\format{\preformatted{ chr ".//|$\{ROOT\}"
}}
\usage{
defaultSearchPaths
}
\description{
The default Search Paths
}
\details{
\itemize{

  \item \dQuote{.//} - the directory containing the file which has
initiated the search

  \item \dQuote{${ROOT}} - the directory from which the glue system
has been invoked }
}
\keyword{datasets}

