% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dir.copy}
\alias{dir.copy}
\title{Copy a directory and its contents}
\usage{
dir.copy(from, to, overwrite = FALSE)
}
\arguments{
\item{from}{directory to copy}

\item{to}{destination directory}

\item{overwrite}{logical; should existing destination files be
overwritten}
}
\value{
logical vector indicating which files were successfully
    copied
}
\description{
Copy a directory and its contents
}
\details{
Copies a directory and all its contents recursively to a
    new location. If the \file{to} directory does not exist it will
    be created. If it does exist the contens of \file{from} will be
    copied into \code{to}. By default contents of \file{to} will
    not be overwritten.
}

