% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerHost.R
\name{dockerHost}
\alias{dockerHost}
\title{Create a \code{dockerHost} object}
\usage{
dockerHost(image, guestdir = "/home/conduit")
}
\arguments{
\item{image}{Name of Docker image}

\item{guestdir}{Guest synced folder path}
}
\value{
Object of class \code{dockerHost} and \code{moduleHost}
}
\description{
Create a \code{moduleHost} object of class \code{dockerHost} to
describe a docker image for executing module source scripts.
}
\details{
The default Docker setup for docker synced folders is that no directories
are shared.  
The default 'conduit' setup for docker synced folders is that the
directory that the module is run within is shared with \file{/home/conduit}
on the guest container.
The synced folder location on the guest container can be customised
with \code{guestdir}
}

