% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{end.pipe}
\alias{end.pipe}
\title{Return \code{pipe} end list}
\usage{
\method{end}{pipe}(x, ...)
}
\arguments{
\item{x}{\code{pipe} object}

\item{...}{additional arguments}
}
\value{
list containing \code{component} and \code{input}
}
\description{
Return \code{pipe} end list
}

