% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{endComponent}
\alias{endComponent}
\title{Return \code{pipe} end component}
\usage{
endComponent(pipe)
}
\arguments{
\item{pipe}{\code{pipe} object}
}
\value{
name of end component as string
}
\description{
Return \code{pipe} end component
}

