% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{endInput}
\alias{endInput}
\title{Return \code{pipe} end input}
\usage{
endInput(pipe)
}
\arguments{
\item{pipe}{\code{pipe} object}
}
\value{
name of end input as string
}
\description{
Return \code{pipe} end input
}

