% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureFileInput}
\alias{ensureFileInput}
\title{ensure input described by fileVessel is satisfied}
\usage{
ensureFileInput(ref, outputObject)
}
\arguments{
\item{ref}{file path to be satisfied}

\item{outputObject}{file path to be passed to ref}
}
\description{
ensure input described by fileVessel is satisfied
}

