% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureFileVessel}
\alias{ensureFileVessel}
\title{ensure input described by fileVessel is satisfied}
\usage{
ensureFileVessel(ref, resource)
}
\arguments{
\item{ref}{file path to be satisfied}

\item{resource}{file path to be passed to ref}
}
\description{
ensure input described by fileVessel is satisfied
}

