% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureInternalInput}
\alias{ensureInternalInput}
\title{ensure input described by internalVessel is satisfied}
\usage{
ensureInternalInput(symbol, resource, language)
}
\arguments{
\item{symbol}{language symbol to be satisfied}

\item{resource}{location of serialised language object}

\item{language}{module script language}
}
\value{
script as character vector
}
\description{
ensure input described by internalVessel is satisfied
}
\seealso{
\code{moduleInputScript}, \code{executeScript}
}

