% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureInternalOutput}
\alias{ensureInternalOutput}
\title{ensure output described by internalVessel is created}
\usage{
ensureInternalOutput(symbol, language)
}
\arguments{
\item{symbol}{language symbol named as output}

\item{language}{module script language}
}
\value{
character vector of script to create serialized output object
}
\description{
ensure output described by internalVessel is created
}
\seealso{
\code{ensureModuleOutput} and \code{executeScript}
}

