% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureInternalVessel}
\alias{ensureInternalVessel}
\title{ensure input described by internalVessel is satisfied}
\usage{
ensureInternalVessel(symbol, resource, language)
}
\arguments{
\item{symbol}{}

\item{resource}{}

\item{language}{}
}
\value{
script as character vector
}
\description{
ensure input described by internalVessel is satisfied
}
\seealso{
\code{moduleInputScript}, \code{executeScript}
}

