% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureModuleInput}
\alias{ensureModuleInput}
\title{ensure module inputs will be satisfied}
\usage{
ensureModuleInput(input, inputObject, language)
}
\arguments{
\item{input}{\code{moduleInput} object}

\item{inputObject}{address of resource to be supplied as input}

\item{language}{language of module script}
}
\value{
character string containing script to ensure input or NULL
}
\description{
ensure module inputs will be satisfied
}

