% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureModuleOutput}
\alias{ensureModuleOutput}
\title{ensure module outputs are produced}
\usage{
ensureModuleOutput(output, language)
}
\arguments{
\item{output}{\code{moduleOutput} object}

\item{language}{language of module script}
}
\value{
character string containing script to produce output or NULL
}
\description{
ensure module outputs are produced
}
\seealso{
\code{executeScript}
}

