% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{execLanguage}
\alias{execLanguage}
\title{Select execLanguage to store in \code{moduleLanguage} object}
\usage{
execLanguage(language, minVersion = NULL, maxVersion = NULL,
  version = NULL)
}
\value{
execLanguage as character
}
\description{
Select execLanguage to store in \code{moduleLanguage} object
}
\details{
For most languages this function will simply return
    \code{language}. For \samp{python} this function returns
    \samp{python2} or \samp{python3} depending on the version
    specifications; the default is \samp{python3}.
}
\seealso{
\code{moduleLanguage}
}

