% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R, R/vagrantHost.R
\name{executeCommand}
\alias{executeCommand}
\alias{executeCommand.default}
\alias{executeCommand.vagrantHost}
\title{Execute a \code{command} list object}
\usage{
executeCommand(moduleHost, hostSubdir, command)

\method{executeCommand}{default}(moduleHost, hostSubdir, command)

\method{executeCommand}{vagrantHost}(moduleHost, hostSubdir, command)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}

\item{hostSubdir}{module output directory on host machine}

\item{command}{\code{command} object}
}
\value{
0 if successful
}
\description{
These methods execute a command list prepared by the \code{command}
function.
}
\details{
If a \code{moduleHost} is provided the command is executed in the
\code{hostSubdir} on the host machine.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: execute a command with no
\code{moduleHost}

\item \code{vagrantHost}: execute command on a \code{vagrantHost}
}}
\seealso{
\code{moduleHost}, \code{prepareModuleHost} for hostSubdir
    creation, \code{command}
}

