% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.R
\name{executeScript.python}
\alias{executeScript.python}
\title{Execute a script in the "python" language}
\usage{
\method{executeScript}{python}(module, inputObjects)
}
\arguments{
\item{module}{\code{module} object}

\item{inputObjects}{Named list of input objects}
}
\value{
named list of \code{moduleOutput} objects
}
\description{
Execute a script in the "python" language
}
\details{
Creates a .py script file from the supplied \code{module},
taking specific input file paths from \code{inputs}.

Writes script to \code{modulePath}, then attempts to execute the
script in this location.
}

