% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.R
\name{executeScript.R}
\alias{executeScript.R}
\title{Execute a script in the "R" language}
\usage{
\method{executeScript}{R}(module, resources)
}
\arguments{
\item{module}{\code{module} object}

\item{resources}{Named list of input objects}
}
\value{
FIXME: nothing meaningful
}
\description{
Execute a script in the "R" language
}
\details{
Creates a .R script file from the supplied \code{module},
taking specific input file paths from \code{inputs}.

Writes script to \code{modulePath}, then attempts to execute the
script in this location.
}

