% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{executeScript}
\alias{executeScript}
\title{Execute a prepared module script file.}
\usage{
executeScript(script)
}
\arguments{
\item{script}{script file to be executed}
}
\value{
named list of \code{moduleOutput} objects
}
\description{
Execute a prepared module script file.
}
\seealso{
\code{runModule}
}

