% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/result.R
\name{export}
\alias{export}
\alias{export.componentResult}
\title{Export OpenAPI objects}
\usage{
export(x, targetDirectory = getwd())

\method{export}{componentResult}(x, targetDirectory = getwd())
}
\arguments{
\item{x}{object}

\item{targetDirectory}{file location}
}
\description{
Export OpenAPI objects for sharing.
}
\details{
This function creates tarball bundles of OpenAPI objects to be
shared and loaded in other sessions.
}
\section{Methods (by class)}{
\itemize{
\item \code{componentResult}: Export a \code{componentResult} object
}}

