% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/component.R
\name{exportComponent}
\alias{exportComponent}
\title{Export a component to an XML file}
\usage{
exportComponent(component, targetDirectory = getwd())
}
\arguments{
\item{component}{\code{component} object}

\item{targetDirectory}{File path for pipeline output}
}
\value{
Resulting file path
}
\description{
Export a component to an XML file
}

