% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{extractModuleSource.fileVessel}
\alias{extractModuleSource.fileVessel}
\title{Extract a module's source script from a fileVessel}
\usage{
\method{extractModuleSource}{fileVessel}(moduleSource)
}
\description{
Extract a module's source script from a fileVessel
}

