% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{extractModuleSource}
\alias{extractModuleSource}
\title{Extract a module's source script}
\usage{
extractModuleSource(moduleSource)
}
\arguments{
\item{vessel}{Vessel indicating source script location}
}
\value{
character vector of source scrip
}
\description{
Extract a module's source script
}
\seealso{
\code{executeScript}
}

