% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{extractModuleSource.scriptVessel}
\alias{extractModuleSource.scriptVessel}
\title{Extract a module's source script from a scriptVessel}
\usage{
\method{extractModuleSource}{scriptVessel}(moduleSource)
}
\description{
Extract a module's source script from a scriptVessel
}

