% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{fetchFromHost}
\alias{fetchFromHost}
\title{Fetch file from remote host.}
\usage{
fetchFromHost(file, host)
}
\arguments{
\item{file}{file path}

\item{host}{host list}
}
\value{
0 if successful
}
\description{
Fetch file from remote host.
}
\details{
copies file at \code{file} from remote directory
\code{host$directory} to the current working directory.
}

