% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/vessel.R
\name{fetchVessel}
\alias{fetchVessel}
\alias{fetchVessel.fileVessel}
\alias{fetchVessel.urlVessel}
\title{Returns the contents of a resource referenced by a \code{vessel} object.}
\usage{
fetchVessel(vessel, location = getwd())

\method{fetchVessel}{fileVessel}(vessel, location = getwd())

\method{fetchVessel}{urlVessel}(vessel, location = getwd())
}
\arguments{
\item{vessel}{\code{vessel} object}

\item{location}{file path of parent object containing vessel}
}
\value{
character vector of file contents

The character vector resulting from a \code{fileVessel}
    object will have an attribute, \code{location}, which contains
    the path to the original file object.
}
\description{
Returns the contents of a resource referenced by a \code{vessel} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{fileVessel}: Return the text of a file resource

\item \code{urlVessel}: Return the text of a URL resource
}}

