% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchXMLAttr}
\alias{fetchXMLAttr}
\title{Fetches an XML attribute if it exists}
\usage{
fetchXMLAttr(xmlNode, attrName)
}
\arguments{
\item{xmlNode}{\code{XMLNode} whose attributes are to be extracted}

\item{attrName}{String vale naming attribute required}
}
\value{
Character string containing attribute value, or NULL
}
\description{
Fetches an XML attribute if it exists
}

