% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{fileIO}
\alias{fileIO}
\title{Create a file IO object}
\usage{
fileIO(ref, path = NULL)
}
\arguments{
\item{ref}{Absolute or relative location of file}

\item{path}{Optional search path for locating file}
}
\description{
Create a file IO object
}
\seealso{
\code{moduleIO}
}

