% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{fileToHost}
\alias{fileToHost}
\title{Copy a file to remote module host}
\usage{
fileToHost(file, host, idfile = get("defaultIdfile", envir = .conduit.global))
}
\arguments{
\item{file}{file to copy}

\item{host}{host list}

\item{idfile}{login credentials}
}
\value{
0 if successful
}
\description{
Copy a file to remote module host
}
\details{
copies file at \code{file} to remote directory
\code{host$directory}.
}
\seealso{
\code{runModule}
}

