% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vessel.R
\name{fileVessel}
\alias{fileVessel}
\title{Create a \code{fileVessel} object.}
\usage{
fileVessel(ref, path = NULL)
}
\arguments{
\item{ref}{Absolute or relative location of file}

\item{path}{Optional search path for locating file}
}
\value{
\code{fileVessel}, \code{vessel} list object
}
\description{
Create a \code{fileVessel} object.
}
\details{
\code{fileVessel} objects are used to point to files on
the local filesystem.

These can be used to specify:

\itemize{
  \item{to where a \code{moduleOutput} is to be saved}
  \item{from where a \code{moduleInput} is to retrieved}
  \item{from where a \code{moduleSource} is to be retrieved}
}

\code{ref} (and \code{path}, if given), must be character vectors of
length 1.
}
\examples{
data_file <- fileVessel(ref="~/myFiles/data.csv")
}
\seealso{
More about \code{vessel} objects, more on \code{fileVessel}
\code{vessel} list objects, overview of \code{module} objects.
}

