% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{findFile}
\alias{findFile}
\title{Find a file referenced by \code{ref} and \code{path}}
\usage{
findFile(ref, path = NULL)
}
\arguments{
\item{ref}{file path or filename}

\item{path}{search path (optional)}
}
\value{
absolute file path as character or NULL
}
\description{
Find a file referenced by \code{ref} and \code{path}
}

