% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R, R/pipeline.R
\name{getComponents}
\alias{getComponents}
\alias{getComponents.pipeline}
\title{S3 generic to extract components from object}
\usage{
getComponents(x)

\method{getComponents}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
S3 generic to extract components from object
}

