% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R, R/module.R, R/pipeline.R
\name{getDescription}
\alias{getDescription}
\alias{getDescription.module}
\alias{getDescription.pipeline}
\title{S3 generic to extract object description}
\usage{
getDescription(x)

\method{getDescription}{module}(x)

\method{getDescription}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
S3 generic to extract object description
}

