% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{getExecLanguageVersion}
\alias{getExecLanguageVersion}
\title{Recovers module execution language information}
\usage{
getExecLanguageVersion(modulePath)
}
\arguments{
\item{modulePath}{module output file directory}
}
\value{
list containing:
\item{\code{execVersion}}{exact version of executing language as character}
\item{\code{failMain}}{\code{TRUE} if language did not meet minVersion}
\item{\code{failMax}}{\code{TRUE} if language did not meet maxVersion}
\item{\code{failExact}}{\code{TRUE} if language did not match version}
}
\description{
Interprets the \file{.languageVersion} file produced when module
sources script is executed.
}
\details{
The \file{.languageVersion} file should contain four lines of text:

\enumerate{
    \item the exact version of the language used for execution
    \item \samp{1} if language did not meet minVersion, else \samp{0}
    \item \samp{1} if language did not meet maxVersion, else \samp{0}
    \item \samp{1} if language did not match version, else \samp{0}
}
}
\seealso{
prepareScriptInit
}

