% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getFormat}
\alias{getFormat}
\alias{getFormat.default}
\title{Retrieve object format}
\usage{
getFormat(x)

\method{getFormat}{default}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object format
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 
}}

