% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/module.R
\name{getLanguage}
\alias{getLanguage}
\alias{getLanguage.default}
\alias{getLanguage.module}
\alias{getLanguage.moduleLanguage}
\alias{getLanguage.output}
\title{Retrieve object language}
\usage{
getLanguage(x)

\method{getLanguage}{default}(x)

\method{getLanguage}{moduleLanguage}(x)

\method{getLanguage}{module}(x)

\method{getLanguage}{output}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object language
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{moduleLanguage}: Return language name as character

\item \code{module}: Returns \code{moduleLanguage} object

\item \code{output}: Get \code{moduleLanguage} object
}}

