% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/generic.R, R/module.R, R/pipeline.R
\name{getLocation}
\alias{getLocation}
\alias{getLocation.module}
\alias{getLocation.pipeline}
\title{Retrieve object location}
\usage{
getLocation(x)

\method{getLocation}{module}(x)

\method{getLocation}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object location
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Returns location of module XML file

\item \code{pipeline}: Returns location of pipeline XML file
}}

