% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R, R/module.R, R/pipeline.R
\name{getName}
\alias{getName}
\alias{getName.component}
\alias{getName.module}
\alias{getName.pipeline}
\title{Extract object name}
\usage{
getName(x)

\method{getName}{module}(x)

\method{getName}{pipeline}(x)

\method{getName}{component}(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract object name
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Returns module name

\item \code{pipeline}: Returns pipeline name

\item \code{component}: Returns component name
}}

