% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R, R/pipeline.R
\name{getPipes}
\alias{getPipes}
\alias{getPipes.pipeline}
\title{S3 generic to extract pipes from object}
\usage{
getPipes(x)

\method{getPipes}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
S3 generic to extract pipes from object
}

