% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getRef}
\alias{getRef}
\alias{getRef.default}
\title{Retrieve object ref}
\usage{
getRef(x)

\method{getRef}{default}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object ref
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 
}}

