% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getResult}
\alias{getResult}
\alias{getResult.default}
\title{Retrieve object result}
\usage{
getResult(x)

\method{getResult}{default}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object result
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 
}}

