% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R, R/generic.R, R/vessel.R
\name{getType.component}
\alias{getType}
\alias{getType.component}
\alias{getType.default}
\alias{getType.vessel}
\title{Retrieve object type}
\usage{
\method{getType}{component}(x)

getType(x)

\method{getType}{default}(x)

\method{getType}{vessel}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object type
}
\section{Methods (by class)}{
\itemize{
\item \code{component}: Returns component type

\item \code{default}: 

\item \code{vessel}: Return the vessel type.
}}

