% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R, R/generic.R
\name{getValue.component}
\alias{getValue}
\alias{getValue.component}
\title{Retrieve object value}
\usage{
\method{getValue}{component}(x)

getValue(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object value
}
\section{Methods (by class)}{
\itemize{
\item \code{component}: Returns module or pipeline object
}}

