% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R, R/generic.R
\name{getVessel.component}
\alias{getVessel}
\alias{getVessel.component}
\title{Retrieve object vessel}
\usage{
\method{getVessel}{component}(x)

getVessel(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object vessel
}
\section{Methods (by class)}{
\itemize{
\item \code{component}: Returns component vessel
}}

