% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{importModule}
\alias{importModule}
\title{Import from a module archive}
\usage{
importModule(tarfile, name)
}
\arguments{
\item{tarfile}{File path to module archive}

\item{name}{Module name}
}
\value{
\code{module} object
}
\description{
Import a module from an exported module tarfile.
}
\details{
This function is used to import modules from gzipped tar archives,
created by the \code{export.componentFile} function. The tarfile to
be imported must be called \file{MODULENAME.tar.gz}; the module to
be loaded from the archive will be called
\code{MODULENAME/MODULENAME.xml}. The resulting \code{module}
object can be named using the \code{name} argument--it will be
called MODULENAME by default.
}
\seealso{
\code{export.componentResult}
}

