% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{importPipeline}
\alias{importPipeline}
\title{Import from a pipeline archive}
\usage{
importPipeline(tarfile, name)
}
\arguments{
\item{tarfile}{File path to pipeline archive}

\item{name}{Pipeline name}
}
\value{
\code{pipeline} object
}
\description{
Import a pipeline from an exported pipeline tarfile.
}
\details{
This function is used to import pipelines from gzipped tar
archives, created by the \code{export.componentFile} function. The
tarfile to be imported must be called \file{PIPELINENAME.tar.gz};
the PIPELINE to be loaded from the archive will be found in the
file \code{PIPELINENAME/pipeline.xml}. The resulting
\code{pipeline} object can be named using the \code{name}
argument--it will be called PIPELINENAME by default.
}
\seealso{
\code{export.componentResult}
}

