% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{input}
\alias{input}
\title{Create an \code{input} object for executing a \code{component} in a
\code{pipeline}.}
\usage{
input(pipe, outputList)
}
\arguments{
\item{pipe}{\code{pipe} describing match}

\item{outputList}{named list of \code{output} object lists}
}
\value{
named input list
}
\description{
Create an \code{input} object for executing a \code{component} in a
\code{pipeline}.
}
\details{
When a pipeline is executed, conduit must match
    \code{moduleOutput}s to \code{moduleInput}s according to the
    \code{pipe}s. This function matches each \code{moduleInput}
    mentioned in a \code{pipe} to the correct \code{output}.

\code{outputList} is a list of \code{output} object lists, named
for each component in the originating pipeline.
}

