% Generated by roxygen2 (4.0.2): do not edit by hand
\name{internalExtension}
\alias{internalExtension}
\title{Returns the correct file extension for a platform's 'internal' files}
\usage{
internalExtension(platform)
}
\arguments{
\item{platform}{platform name}
}
\value{
files exension as character as ".EXT"
}
\description{
Returns the correct file extension for a platform's 'internal' files
}

