% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalInputScript.python}
\alias{internalInputScript.python}
\title{prepare internal input script for python language}
\usage{
\method{internalInputScript}{python}(symbol, resource)
}
\description{
prepare internal input script for python language
}

