% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.R
\name{internalInputScript.R}
\alias{internalInputScript.R}
\title{prepare internal input script for R language}
\usage{
\method{internalInputScript}{R}(symbol, resource)
}
\description{
prepare internal input script for R language
}

