% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shell.R
\name{internalInputScript.shell}
\alias{internalInputScript.shell}
\title{prepare internal input script for shell language}
\usage{
\method{internalInputScript}{shell}(symbol, inputObject)
}
\description{
prepare internal input script for shell language
}

