% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{internalIO}
\alias{internalIO}
\title{Create an internal IO object}
\usage{
internalIO(symbol)
}
\arguments{
\item{symbol}{Name of symbol identifier}
}
\value{
named \code{internalIO} list item
}
\description{
Create an internal IO object
}
\seealso{
\code{moduleInput}
}

