% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.R
\name{internalOutputScript.python}
\alias{internalOutputScript.python}
\title{create script to create internal output for language = "python"}
\usage{
\method{internalOutputScript}{python}(symbol)
}
\description{
create script to create internal output for language = "python"
}

