% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.R
\name{internalOutputScript.R}
\alias{internalOutputScript.R}
\title{create script to create internal output for language = "R"}
\usage{
\method{internalOutputScript}{R}(symbol)
}
\description{
create script to create internal output for language = "R"
}

