% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shell.R
\name{internalOutputScript.shell}
\alias{internalOutputScript.shell}
\title{create script to create internal output for language = "shell"}
\usage{
\method{internalOutputScript}{shell}(symbol)
}
\description{
create script to create internal output for language = "shell"
}

