% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vessel.R
\name{internalVessel}
\alias{internalVessel}
\title{Create an \code{internalVessel} object.}
\usage{
internalVessel(symbol)
}
\arguments{
\item{symbol}{Name of internal object}
}
\value{
\code{internalVessel}, \code{vessel} list object
}
\description{
Create an \code{internalVessel} object.
}
\details{
\code{internalVessel} objects are used to point to objects
internal to a \code{module} script. They are used to pass objects
internal to a module's platform to other modules.

These can be used to specify:

\itemize{
  \item{an object created in a \code{module} script to be passed out as
        a \code{moduleOutput}}
  \item{an object which a \code{module} script expects to be provided
        by a \code{moduleInput}}
}

\code{symbol} must be a character vectors of length 1.
}
\examples{
expenses_df <- internalVessel(symbol="expenses")
}
\seealso{
More about \code{vessel} objects, more about
\code{moduleInput} and \code{moduleOutput} objects, overview of
\code{module} objects.
}

