% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalVesselScript.python}
\alias{internalVesselScript.python}
\title{prepare internalVessel input script for python language}
\usage{
\method{internalVesselScript}{python}(symbol, resource)
}
\description{
prepare internalVessel input script for python language
}

