% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalVesselScript.R}
\alias{internalVesselScript.R}
\title{prepare internalVessel input script for R language}
\usage{
\method{internalVesselScript}{R}(symbol, resource)
}
\description{
prepare internalVessel input script for R language
}

