% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalVesselScript}
\alias{internalVesselScript}
\title{prepare internalVessel input script}
\usage{
internalVesselScript(symbol, resource)
}
\arguments{
\item{symbol}{character string with class set to language of module script}

\item{resource}{file location of serialised language object}
}
\value{
character vector of script to ensure input
}
\description{
prepare internalVessel input script
}

