% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{ioFormatToXML}
\alias{ioFormatToXML}
\title{Create XML corresponding to an \code{ioFormat} object}
\usage{
ioFormatToXML(ioFormat, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{ioFormat}{\code{ioFormat} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to an \code{ioFormat} object
}
\seealso{
\code{ioFormat} objects, \code{XML::newXMLNode}
}

