% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_absolute}
\alias{is_absolute}
\title{Tests if a file path is absolute}
\usage{
is_absolute(path)
}
\arguments{
\item{path}{file path}
}
\value{
TRUE if absolute
}
\description{
Tests if a file path is absolute
}

