% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{is_length1_char}
\alias{is_length1_char}
\title{Checks whether \code{x} is a length 1 character vector}
\usage{
is_length1_char(x)
}
\arguments{
\item{x}{vector to be checked}
}
\value{
logical, TRUE if \code{x} is length 1 char
}
\description{
Checks whether \code{x} is a length 1 character vector
}

