% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadModule}
\alias{loadModule}
\title{Load a module from an XML file}
\usage{
loadModule(name, ref, path = NULL, namespaces = c(oa =
  "http://www.openapi.org/2014/"))
}
\arguments{
\item{name}{Name of module}

\item{ref}{Module location or filename}

\item{path}{(Optional) Search path if \code{ref} is a filename}

\item{namespaces}{Namespaces used in XML document}
}
\value{
\code{module} list
}
\description{
Load a module from an XML file
}
\seealso{
\code{module}
}

