% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadPipeline}
\alias{loadPipeline}
\title{Load a pipeline from an XML file}
\usage{
loadPipeline(name, ref, path = NULL, namespaces = c(oa =
  "http://www.openapi.org/2014/"))
}
\arguments{
\item{name}{Name of pipeline}

\item{ref}{Path to XML file}

\item{path}{Search path (optional)}

\item{namespaces}{Namespaces used in XML document}
}
\value{
\code{pipeline} list
}
\description{
Load a pipeline from an XML file
}

