% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{matchInput}
\alias{matchInput}
\title{Match a pipe's input name to an output object}
\usage{
matchInput(pipe, outputObjects)
}
\arguments{
\item{pipe}{\code{pipe} describing match}

\item{outputObjects}{named list of output objects}
}
\value{
named input list
}
\description{
Match a pipe's input name to an output object
}

