% Generated by roxygen2 (4.0.1): do not edit by hand
\name{module}
\alias{module}
\title{Create a module}
\usage{
module(name, description = "", platform, inputs = list(),
  outputs = list(), sources = list(), path = NULL)
}
\arguments{
\item{name}{Name of module as character value}

\item{description}{Description of what module does as character value}

\item{platform}{Module platform as character value}

\item{inputs}{List of \code{moduleInput} objects}

\item{outputs}{List of \code{moduleOutput} objects}

\item{sources}{List of \code{moduleSource} objects}

\item{path}{Search path for original module XML file}
}
\description{
Create a \code{module} list object
}

