% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\name{moduleHost}
\alias{moduleHost}
\alias{vagrantHost}
\title{Constructor functions for \code{moduleHost} objects}
\usage{
vagrantHost(vagrantfile, location = NULL)
}
\arguments{
\item{vagrantfile}{Path to vagrantfile}

\item{location}{File location for running vagrant machine}
}
\value{
Object of class \code{moduleHost}, and one of:

\item{\code{vagrantHost}}{containing \code{vagrantfile} and
    \code{location}}
}
\description{
These functions are used to create \code{moduleHost} objects.
}
\section{Functions}{
\itemize{
\item \code{vagrantHost}: Create a \code{vagrantHost} object
}}

