% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R, R/vagrantHost.R
\name{moduleHostToXML}
\alias{moduleHostToXML}
\alias{moduleHostToXML.vagrantHost}
\title{Create XML corresponding to a \code{moduleHost} object}
\usage{
moduleHostToXML(moduleHost)

\method{moduleHostToXML}{vagrantHost}(vagrantHost)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}
}
\value{
XML node representing module host
}
\description{
Create XML corresponding to a \code{moduleHost} object
}
\section{Methods (by class)}{
\itemize{
\item \code{vagrantHost}: Create XML corresponding to a
\code{vagrantHost} object
}}

