% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vagrantHost.R
\name{moduleHostToXML.vagrantHost}
\alias{moduleHostToXML.vagrantHost}
\title{Create XML corresponding to a \code{vagrantHost} object}
\usage{
\method{moduleHostToXML}{vagrantHost}(vagrantHost)
}
\description{
Create XML corresponding to a \code{vagrantHost} object
}

