% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleInput}
\alias{moduleInput}
\title{Create a \code{module} input}
\arguments{
\item{name}{Input name}

\item{type}{Input type}

\item{...}{attributes appropriate to \code{type}}

\item{format}{Input format}

\item{formatType}{Defaults to \dQuote{text}}
}
\value{
named \code{moduleInput} list containing:
\itemize{
  \item{name}
  \item{type}
  \item{attributes}
  \item{format}
  \item{formatType}
}
}
\description{
Creates a \code{moduleInput} list for use in a \code{module}'s inputs list.
}
\details{
Creates a \code{moduleInput} list with attributes appropriate to
\code{type}. Attributes given by \code{...} argument.

The following \code{type}s and attributes are accepted:
\itemize{
  \item{
    internal:
    \itemize{
      \item{symbol: symbol name}
    }
  }
  \item{
    file
    \itemize{
      \item{ref: relative or absolute path name to file}
      \item{path: search path to find ref}
    }
  }
}
}
\examples{
internalInput <- moduleInput(name = "bigData", type = "internal",
                             symbol = "bigData", format = "R data frame")
fileInput <- moduleInput(name = "scores.csv", type = "file",
                         ref = "2015-03-13-scores.csv",
                         format = "CSV file")
}
\seealso{
\code{module}
}

