% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInputHost.R
\name{moduleInputHost}
\alias{moduleInputHost}
\title{Create a \code{moduleInputHost} object}
\usage{
moduleInputHost(name)
}
\arguments{
\item{name}{Name of \code{moduleInput}}
}
\value{
Object of class \code{moduleInputHost} and \code{moduleHost}
}
\description{
Create a \code{moduleHost} object of class \code{moduleInputHost}
to indicate that a \code{moduleInput} will provide the real
\code{moduleHost} at execution.
}

