% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleIO}
\alias{moduleIO}
\title{Create a \code{module} input or output object.}
\usage{
moduleIO(name, type, vessel, format)
}
\arguments{
\item{name}{Input/output name}

\item{type}{\dQuote{input} or \dQuote{output}}

\item{vessel}{\code{vessel} object}

\item{format}{\code{ioFormat} object}
}
\value{
\code{moduleIO} object
}
\description{
Create a \code{module} input or output object.
}
\details{
This function is used by \code{moduleInput} and
\code{moduleOutput} to create input and output objects for
modules.

\code{vessel} cannot be a \code{scriptVessel} object, as these are not
defined for moduleIO objects.
}
\seealso{
\code{module}, \code{moduleInput}, \code{moduleOutput},
\code{ioFormat}, \code{vessel}
}

