% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleIOToXML}
\alias{moduleIOToXML}
\title{Create XML corresponding to \code{moduleIO} object}
\usage{
moduleIOToXML(moduleIO, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{moduleIO}{\code{moduleIO} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to \code{moduleIO} object
}
\seealso{
\code{moduleIO} objects, \code{XML::newXMLNode}
}

