% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleLanguage}
\alias{moduleLanguage}
\title{Create a \code{moduleLanguage} object}
\usage{
moduleLanguage(language, minVersion = NULL, maxVersion = NULL,
  version = NULL)
}
\arguments{
\item{language}{Language name}

\item{minVersion}{Minimum version required}

\item{maxVersion}{Maximum version required}

\item{version}{Exact version required}
}
\value{
\code{moduleLanguage} object
}
\description{
Create a \code{moduleLanguage} object
}

