% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleLanguageToXML}
\alias{moduleLanguageToXML}
\title{Create XML corresponding to \code{moduleLanguage} object}
\usage{
moduleLanguageToXML(moduleLanguage)
}
\arguments{
\item{moduleLanguage}{\code{moduleLanguage} objects}
}
\value{
XML node representing module language
}
\description{
Create XML corresponding to \code{moduleLanguage} object
}

