% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleOutput}
\alias{moduleOutput}
\title{Create a \code{module} output input}
\usage{
moduleOutput(name, type, format = "", formatType = "text", ref = "")
}
\arguments{
\item{name}{Output name}

\item{type}{\sQuote{internal} or \sQuote{external}}

\item{format}{Output format}

\item{formatType}{Defaults to \dQuote{text}}

\item{ref}{Filename of \sQuote{external} output}
}
\value{
named \code{moduleOutput} character vector of:
\itemize{
  \item{name}
  \item{type}
  \item{format}
  \item{formatType}
  \item{ref}
}
}
\description{
Creates a \code{moduleOutput} vector for use in a \code{module}'s outputs
list.
}
\details{
\code{type} must be \sQuote{internal} or \sQuote{external}.

It \code{type} is \dQuote{external}, a \code{ref} is required. This needs
to be a resolveable URI created by the \code{module}'s source(s).
}
\examples{
outp1 <- moduleOutput(name = "bigData", type = "internal",
                      format = "R data frame")
outp2 <- moduleOutput(name = "mediumData", type = "external",
                      format = "CSV file", ref = "mediumData.csv")
}
\seealso{
\code{module}
}

