% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{modulePlatform}
\alias{modulePlatform}
\title{Create a \code{module} platform node}
\usage{
modulePlatform(name)
}
\arguments{
\item{name}{Name of platform}
}
\value{
A named character vector containing the platform name
}
\description{
Create a \code{module} platform node
}
\seealso{
\code{module}
}

