% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{moduleResult}
\alias{moduleResult}
\title{Create \code{moduleResult} object}
\usage{
moduleResult(outputList, modulePath, module)
}
\arguments{
\item{outputList}{list of \code{output} objects}

\item{modulePath}{file path to module output}

\item{module}{\code{module} object which produced \code{outputList}}
}
\value{
object of class \code{componentResult} and
    \code{moduleResult} containing:

\item{name}{module name}
\item{file}{file path to resulting module XML}
\item{component}{resulting \code{module} object}
\item{outputList}{list of \code{output} objects produced by module}
\item{execLanguageVersion}{executing language information}
}
\description{
Create \code{moduleResult} object
}
\details{
This functions is used to construct the objects returned
    by \code{runModule}.

The \code{module} object returned, and its corresponding XML
\code{file}, will specify the version of the language used to
execute the source scripts in its \code{moduleLanguage}.

The \code{export} function can be used to export these objects to a
gzipped tarfile. The resulting tarfile can be read loaded into
conduit using the \code{importModule} function
}
\seealso{
\code{pipelineResult}, \code{module}, \code{runModule},
    \code{export}, \code{importModule}
}

