% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleSource}
\alias{moduleSource}
\title{Create a \code{module} source node}
\usage{
moduleSource(value, ref = NULL, path = defaultSearchPaths, type = "",
  order = "")
}
\arguments{
\item{value}{source script}

\item{ref}{module XML filename}

\item{path}{search path(s)}

\item{type}{\dQuote{module} or \dQuote{pipeline}}

\item{order}{numeric, specifies source position in sources}
}
\description{
Create a \code{module} source node
}
\seealso{
\code{module}
}

