% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleSourceToXML}
\alias{moduleSourceToXML}
\title{Create XML corresponding to \code{moduleSource} object}
\usage{
moduleSourceToXML(moduleSource, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{moduleSource}{\code{moduleSource} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to \code{moduleSource} object
}
\seealso{
\code{XML::newXMLNode}

\code{moduleSource} objects
}

