% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleToXML}
\alias{moduleToXML}
\title{Convert a \code{module} object to XML}
\usage{
moduleToXML(module, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{module}{\code{module} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLInternalNode} object
}
\description{
Convert a \code{module} object to XML
}
\seealso{
\code{module} objects, \code{XML::newXMLNode}
}

