% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{output}
\alias{output}
\title{return \code{output} produced by a \code{moduleOutput}}
\usage{
output(moduleOutput, language, outputDirectory)
}
\arguments{
\item{language}{module script language}

\item{outputDirectory}{file location for module execution}

\item{output}{\code{moduleOutput} object}
}
\value{
\code{output} object
}
\description{
return \code{output} produced by a \code{moduleOutput}
}

