% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/component.R
\name{outputObject}
\alias{outputObject}
\title{return output object produced by a module output}
\usage{
outputObject(output, outputDirectory)
}
\arguments{
\item{output}{\code{moduleOutput} object}

\item{outputDirectory}{file location for module execution}
}
\value{
output object
}
\description{
return output object produced by a module output
}

