% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{parseModuleHost}
\alias{parseModuleHost}
\title{Parse a module's host}
\usage{
parseModuleHost(host)
}
\arguments{
\item{host}{module host as character string}
}
\value{
list of user, address, port
}
\description{
Parse a module's host
}
\details{
Default user = 'conduit', default port = '22'.
}

