% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{pathSep}
\alias{pathSep}
\title{Character used to separate search paths}
\format{\preformatted{ chr "|"
}}
\usage{
pathSep
}
\description{
Character used to separate search paths
}
\seealso{
\code{resolveRef}
}
\keyword{datasets}

