% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pipe}
\alias{pipe}
\title{Create a pipe object}
\usage{
pipe(startComponent, startOutput, endComponent, endInput)
}
\arguments{
\item{startComponent}{Name of start module}

\item{startOutput}{Name of start object}

\item{endComponent}{Name of end module}

\item{endInput}{Name of end input}
}
\value{
\code{pipe} connecting \code{startComponentName}.\code{startOutputName} to \code{endComponentName}.\code{endInputName}
}
\description{
Create a pipe object
}

