% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pipeline}
\alias{pipeline}
\title{Create a pipeline}
\usage{
pipeline(name, path = NULL, description = "", components = list(),
  modules = list(), pipelines = list(), pipes = list())
}
\arguments{
\item{name}{\code{pipeline} name}

\item{path}{location of originating pipeline xml file}

\item{description}{\code{pipeline} description}

\item{components}{list of \code{module}s and \code{pipeline}s}

\item{modules}{list of \code{module}s}

\item{pipelines}{list of \code{pipeline}s}

\item{pipes}{list of \code{pipe}s}
}
\value{
\code{pipeline} list containing:
\item{name}{character value}
\item{path}{Location of source pipeline XML file}
\item{description}{character value}
\item{components}{list of \code{module}s and \code{pipeline}s}
\item{pipes}{list of \code{pipe}s}
}
\description{
Create an openapi \code{pipeline} object
}
\details{
If \code{components} is not specified, \code{pipeline} will construct
it the compoments from \code{modules} and \code{pipelines}.
}

