% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{pipelineResult}
\alias{pipelineResult}
\title{Create \code{pipelineResult} object}
\usage{
pipelineResult(componentResultList, pipelinePath, pipeline)
}
\arguments{
\item{componentResultList}{list of \code{componentResult} objects}

\item{pipelinePath}{file path to pipeline output}

\item{pipeline}{\code{pipeline} object which produced
\code{componentResultList}}
}
\value{
object of class \code{componentResult} and
    \code{pipelineResult} containing:

\item{name}{module name}
\item{file}{file path to resulting pipeline XML}
\item{component}{resulting \code{pipeline} object}
\item{outputList}{list of \code{output} objects produced by components}
\item{componentResultList}{list of \code{componentResult} objects}
}
\description{
Create \code{pipelineResult} object
}
\details{
This function is used to construct the objects returned
    by \code{runPipeline}.

The \code{export} function can be used to export these objects to a
gzipped tarfile. The resulting tarfile can be read loaded into
conduit using the \code{importPipeline} function
}
\seealso{
\code{moduleResult}, \code{pipeline}, \code{runPipeline},
    \code{export}, \code{importPipeline}
}

