% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipesAsEdges}
\alias{pipesAsEdges}
\title{Get pipes as list of edges}
\usage{
pipesAsEdges(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} list object}
}
\value{
named list of character vectors for each component
}
\description{
For each vertex (component) get a vector of vertices to which there
is an edge (pipe).
}

