% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipeToXML}
\alias{pipeToXML}
\title{Convert a pipe to XML}
\usage{
pipeToXML(pipe, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{pipe}{\code{pipe} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLNode} object
}
\description{
Convert a pipe to XML
}
\seealso{
\code{XML::newXMLNode}
}

