% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{prepareFileInput}
\alias{prepareFileInput}
\title{Prepare file input object}
\usage{
prepareFileInput(vessel, inputObject, outputDirectory, location)
}
\arguments{
\item{vessel}{Module input vessel object}

\item{inputObject}{File path to input}

\item{outputDirectory}{File path to module working directory}

\item{location}{File path to originating module XML file}
}
\value{
File path to input file
}
\description{
This function makes sure a module's file input is available to the
module.
}
\details{
If \code{input} is NULL the module is assumed to be
\dQuote{starting} from a file, and the file referenced in
\code{vessel} is returned.

Where \code{input} is not null, and \code{vessel} describes an
absolute path to a file this path is returned. If \code{vessel}
describes a relative path to a file, this file is copied into the
module's outputDirectory, and the resulting file is returned.
}

