% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{prepareInternalInput}
\alias{prepareInternalInput}
\title{Prepare internal input object.}
\usage{
prepareInternalInput(input, symbol, language, outputDirectory)
}
\arguments{
\item{input}{File path to serialized object}

\item{symbol}{Name of module input}

\item{language}{Module language}

\item{outputDirectory}{File path to module working directory}
}
\value{
File path to serialized internal input.
}
\description{
This function makes a serialized internal module input available in
a module's working directory.
}

