% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerHost.R, R/module.R, R/vagrantHost.R
\name{prepareModuleHost.dockerHost}
\alias{prepareModuleHost}
\alias{prepareModuleHost.dockerHost}
\alias{prepareModuleHost.vagrantHost}
\title{Prepare a module host for execution.}
\usage{
\method{prepareModuleHost}{dockerHost}(moduleHost, moduleName, modulePath)

prepareModuleHost(moduleHost, moduleName, modulePath)

\method{prepareModuleHost}{vagrantHost}(moduleHost, moduleName, modulePath)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}

\item{moduleName}{module name}

\item{modulePath}{module output directory}
}
\value{
\code{outputLocation} object
}
\description{
These methods ensure that a \code{moduleHost} will have all
resources required to execute a \code{module}'s source scripts.
}
\details{
These methods return the path to the module output directory as an
\code{outputLocation} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{dockerHost}: prepare \code{dockerHost}. Returns
empty string.

\item \code{vagrantHost}: prepare \code{vagrantHost}; returns
path to unique module output directory relative to hostdir/guestdir
}}

