% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScriptInput}
\alias{prepareScriptInput}
\title{Prepare script to create inputs}
\usage{
prepareScriptInput(input, language)
}
\arguments{
\item{input}{input name}

\item{language}{module language}
}
\value{
Script as character vector
}
\description{
Prepare script to create inputs
}

