% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScriptOutput}
\alias{prepareScriptOutput}
\title{Prepare script to create outputs}
\usage{
prepareScriptOutput(moduleOutput, language)
}
\arguments{
\item{moduleOutput}{\code{moduleOutput} object}

\item{language}{module language}
}
\value{
Script as character vector
}
\description{
Prepare script to create outputs
}
\details{
if a module output is passed to conduit via an
    internalVessel the module source scripts must serialize the
    object after execution. other vessel types do not need this to
    be done by the glue system.
}

