% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{prepareURLInput}
\alias{prepareURLInput}
\title{Prepare URL input object}
\usage{
prepareURLInput(vessel, input)
}
\arguments{
\item{vessel}{Module input vessel object}

\item{input}{URL string}
}
\value{
URL to input resource
}
\description{
This function ensure's a module's URL input is available to the
module.
}
\details{
If \code{input} is NULL the module is assumed to be
\dQuote{starting} from a URL, and the URL referenced in
\code{vessel} is returned.
}

