% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerHost.R
\name{readDockerHostXML}
\alias{readDockerHostXML}
\title{Create a \code{dockerHost} object from <host><docker/></host> XML}
\usage{
readDockerHostXML(dockerHostXML)
}
\arguments{
\item{dockerHostXML}{docker host XML node}
}
\value{
\code{dockerHost} object
}
\description{
Create a \code{dockerHost} object from <host><docker/></host> XML
}
\seealso{
\code{readModuleHostXML}
}

