% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{readIOFormatXML}
\alias{readIOFormatXML}
\title{Create a \code{ioFormat} object from format XML}
\usage{
readIOFormatXML(xml)
}
\arguments{
\item{xml}{format XML}
}
\value{
\code{ioFormat} object
}
\description{
Create a \code{ioFormat} object from format XML
}

