% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleHostXML}
\alias{readModuleHostXML}
\title{Create a \code{moduleHost} object from host XML}
\usage{
readModuleHostXML(moduleHostXML)
}
\arguments{
\item{moduleHostXML}{host XML}
}
\value{
\code{moduleHost} object
}
\description{
This function creates a \code{moduleHost} object from valid host
elements.
}
\details{
As of 2016-02-26 only \samp{<vagrant/>} elements are
supported.
}

