% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInputHost.R
\name{readModuleInputHostXML}
\alias{readModuleInputHostXML}
\title{Create a \code{moduleInputHost} object from <host><moduleInput/></host> XML}
\usage{
readModuleInputHostXML(moduleInputHostXML)
}
\arguments{
\item{moduleInputHostXML}{moduleInput host XML node}
}
\value{
\code{moduleInputHost} object
}
\description{
Create a \code{moduleInputHost} object from <host><moduleInput/></host> XML
}
\seealso{
\code{readModuleHostXML}
}

