% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleLanguageXML}
\alias{readModuleLanguageXML}
\title{Create a \code{moduleLanguage} object from language XML}
\usage{
readModuleLanguageXML(moduleLanguageXML)
}
\arguments{
\item{moduleLanguageXML}{language XML node}
}
\value{
\code{moduleLanguage} object
}
\description{
Create a \code{moduleLanguage} object from language XML
}

