% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleSourceXML}
\alias{readModuleSourceXML}
\title{create a \code{moduleSource} object from module source XML}
\usage{
readModuleSourceXML(xml)
}
\arguments{
\item{xml}{module source XML}
}
\value{
\code{moduleSource} object
}
\description{
create a \code{moduleSource} object from module source XML
}

