% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{readPipeXML}
\alias{readPipeXML}
\title{Create \code{pipe} object from pipe XML}
\usage{
readPipeXML(node)
}
\arguments{
\item{node}{pipe node}
}
\value{
\code{pipe} object
}
\description{
Create \code{pipe} object from pipe XML
}

