% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\name{readVagrantHostXML}
\alias{readVagrantHostXML}
\title{Create a \code{vagrantHost} object from <host><vagrant/></host> XML}
\usage{
readVagrantHostXML(vagrantHostXML)
}
\arguments{
\item{vagrantHostXML}{vagrant host XML node}
}
\value{
\code{vagrantHost} object
}
\description{
Create a \code{vagrantHost} object from <host><vagrant/></host> XML
}
\seealso{
\code{readModuleHostXML}
}

