% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{readVesselXML}
\alias{readVesselXML}
\title{Create a \code{vessel} object from vessel XML}
\usage{
readVesselXML(xml)
}
\arguments{
\item{xml}{vessel XML}
}
\value{
\code{vessel} object
}
\description{
Create a \code{vessel} object from vessel XML
}

