% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{resolveInput.internal}
\alias{resolveInput.internal}
\title{Resolve internal input object}
\usage{
\method{resolveInput}{internal}(moduleInput, inputObjects)
}
\description{
Resolve internal input object
}

