% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{resolveInput.internal}
\alias{resolveInput}
\alias{resolveInput.file}
\alias{resolveInput.internal}
\title{Resolve input object}
\usage{
\method{resolveInput}{internal}(moduleInput, inputObjects, host)

\method{resolveInput}{file}(moduleInput, inputObjects, host)

resolveInput(moduleInput, inputObjects, host)
}
\arguments{
\item{moduleInput}{\code{moduleInput} object}

\item{inputObjects}{resources to be supplied as inputs}

\item{host}{module host}
}
\value{
boolean
}
\description{
This function ensures that an input object is where it needs to be
for a module's source script to be executed. Returns TRUE if all is well.
}
\details{
If the \code{moduleInput}'s vessel is a \code{fileVessel}
containing a relative 'ref' the inputObject is copied to the
current working directory as 'ref'.
}
\section{Methods (by class)}{
\itemize{
\item \code{internal}: Resolve internal input object

\item \code{file}: Resolve file input object
}}

