% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{resolveInputObject}
\alias{resolveInputObject}
\title{Resolve input object}
\usage{
resolveInputObject(input, inputObject)
}
\arguments{
\item{input}{\code{moduleInput} object}

\item{inputObject}{resource to be supplied as input}
}
\value{
boolean
}
\description{
This function ensures that an input object is where it needs to be
for a module's source script to be executed. Returns TRUE if all is well.
}
\details{
If the \code{moduleInput}'s vessel is a \code{fileVessel}
containing a relative 'ref' the inputObject is copied to the
current working directory as 'ref'.
}

