% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{resolveRef}
\alias{resolveRef}
\title{Resolves a full path for a given ref and path}
\usage{
resolveRef(ref, path = NULL, location = getwd())
}
\arguments{
\item{ref}{address/filename of referenced file}

\item{path}{file paths to search}

\item{location}{file directory of invoking pipeline/module xml}
}
\value{
character string of resolved ref with class set to
appropriate read method
}
\description{
Resolves a full path for a given ref and path
}

