% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{resultComponent}
\alias{resultComponent}
\title{Return a \code{component} object for use in a \code{pipelineResult}
object.}
\usage{
resultComponent(componentResult, pipelinePath)
}
\arguments{
\item{componentResult}{\code{componentResult} object}

\item{pipelinePath}{file path to pipeline output}
}
\value{
\code{component} object
}
\description{
Return a \code{component} object for use in a \code{pipelineResult}
object.
}

