% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{resultInput}
\alias{resultInput}
\title{Return a \code{moduleInput} object for use in a
\code{moduleResult} object.}
\usage{
resultInput(output, modulePath)
}
\arguments{
\item{output}{\code{output} object}

\item{modulePath}{file path to module output}
}
\value{
\code{moduleInput} object or NULL
}
\description{
This function constructs an appropriate \code{moduleInput} object
from the \code{output} object created when executing a module
source.
}

