% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{resultOutput}
\alias{resultOutput}
\title{Return a \code{moduleOutput} object for use in a
\code{moduleResult} object.}
\usage{
resultOutput(output)
}
\arguments{
\item{output}{\code{output} object}
}
\value{
\code{moduleOutput} object
}
\description{
Return a \code{moduleOutput} object for use in a
\code{moduleResult} object.
}

