% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R, R/vagrantHost.R
\name{retrieveModuleHost}
\alias{retrieveModuleHost}
\alias{retrieveModuleHost.vagrantHost}
\title{Retrieve results of running a module's source scripts on a
\code{moduleHost}}
\usage{
retrieveModuleHost(moduleHost, outputLocation, modulePath)

\method{retrieveModuleHost}{vagrantHost}(moduleHost, outputLocation, modulePath)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}

\item{outputLocation}{\code{outputLocation} object}

\item{modulePath}{output directory on local machine}
}
\description{
This functions retrieves the results of executing a module's source
scripts on a \code{moduleHost} using \code{executeScript}.
}
\details{
\code{outputLocation} should be the result of
\code{prepareModuleHost} on the same \code{moduleHost}.
}
\section{Methods (by class)}{
\itemize{
\item \code{vagrantHost}: retrieve module output from
\code{vagrantHost}
}}
\seealso{
\code{moduleHost}, \code{prepareModuleHost},
    \code{executeScript}
}

