% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/component.R
\name{runComponent}
\alias{runComponent}
\title{Run a component}
\usage{
runComponent(componentName, pipeline, inputs = list(),
  pipelinePath = getwd())
}
\arguments{
\item{componentName}{Name of component to be executed}

\item{pipeline}{\code{pipeline} containing component}

\item{inputs}{Named list of absolute paths for component inputs}

\item{pipelinePath}{Pipeline output directory}
}
\value{
FIXME: Result of \code{runModule} or \code{runPipeline}
}
\description{
This function executes a single pipeline component with \code{runModule} or
\code{runPipeline}.
}
\details{
If the component refers to a \code{module}, the names of
\code{inputs} must match the names the module's input names.

\code{pipelinePath} must exist on the filesystem.
}
\examples{
mod1 <- module("setX", platform="R",
               sources=list(moduleSource("x <- \\"set\\"")))
pip1 <- pipeline("setX-pipe", description="set the value of x",
                 components=list(mod1))
runComponent(componentName = "setX", pipeline = pip1)
}
\seealso{
`runPipeline` and `runModule`
}

