% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runModule}
\alias{runModule}
\title{Execute a \code{module}'s \code{moduleSource}s.}
\usage{
runModule(module, inputs = list(), targetDirectory = getwd())
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Named list of input locations}

\item{moduleFiles}{File path for module output}
}
\description{
Execute a \code{module}'s \code{moduleSource}s.
}
\details{
This function:
\itemize{
  \item creates a directory for the \code{module} output
  \item determines which platform the module requires
  \item executes the \code{module}'s \code{moduleSource}s in the specified
        platform
}
}
\seealso{
\code{module}, \code{moduleSource}
}

