% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/fileSystem.R
\name{runPlatform.fileSystem}
\alias{runPlatform.fileSystem}
\title{Platform support for "fileSystem" platform.}
\usage{
runPlatform.fileSystem(module, inputs, modulePath)
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Named list of input locations}

\item{modulePath}{File path for module output}
}
\value{
FIXME: nothing meaningful
}
\description{
Platform support for "fileSystem" platform.
}
\details{
This platform support function is a wrapper which allows
files to be retrieved from the file system and passed as an
\sQuote{external} object to other modules.
}

