% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/python3.R
\name{runPlatform.python3}
\alias{runPlatform.python3}
\title{Platform support for "python3" platform.}
\usage{
runPlatform.python3(module, inputs, modulePath)
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Named list of input locations}

\item{modulePath}{File path for module output}
}
\value{
FIXME: nothing meaningful
}
\description{
Platform support for "python3" platform.
}
\details{
Creates a .py script file from the supplied \code{module},
taking specific input file paths from \code{inputs}.

Writes script to \code{modulePath}, then attempts to execute the
script in this location.
}

