% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runPlatform.R}
\alias{runPlatform.R}
\title{Platform support for "R" platform.}
\usage{
\method{runPlatform}{R}(module, inputs, modulePath)
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Named list of input locations}

\item{modulePath}{File path for module output}
}
\description{
Platform support for "R" platform.
}
\details{
Creates a .R script file from the supplied \code{module},
taking specific input file paths from \code{inputs}.

Writes script to \code{modulePath}, then attempts to execute the
script in this location.
}

