% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{runPlatform}
\alias{runPlatform}
\title{Execute a \code{module}'s \code{moduleSource}s in the specified
platform.}
\usage{
runPlatform(module, inputs, moduleFiles)
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Names list of input locations}

\item{moduleFiles}{File path to module output location}
}
\description{
Execute a \code{module}'s \code{moduleSource}s in the specified
platform.
}

