% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveModule}
\alias{saveModule}
\title{Save a module to disk}
\usage{
saveModule(module, targetDirectory = getwd(), filename = paste0(module$name,
  ".xml"))
}
\arguments{
\item{module}{\code{module} object}

\item{targetDirectory}{destination directory as file path}

\item{filename}{character string of desired filename}
}
\value{
character value of resulting file location
}
\description{
Save a \code{module} to an XML file on disk
}
\details{
The resulting XML file will be called \file{MODULE_NAME.xml} unless another
\code{filename} is specified.
}

