% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{scriptExtension}
\alias{scriptExtension}
\title{Returns the correct file extension for a language's script files}
\usage{
scriptExtension(moduleLanguage)
}
\arguments{
\item{moduleLanguage}{\code{moduleLanguage} object}
}
\value{
file extension as character ".EXT"
}
\description{
Returns the correct file extension for a language's script files
}

