% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vessel.R
\name{scriptVessel}
\alias{scriptVessel}
\title{Create a \code{scriptVessel} object.}
\usage{
scriptVessel(value)
}
\arguments{
\item{value}{Character vector of module script}
}
\value{
\code{scriptVessel}, \code{vessel} list object
}
\description{
Create a \code{scriptVessel} object.
}
\details{
\code{scriptVessel} objects are used to store inline module
source scripts.
}
\examples{
source2_script <-
    scriptVessel(
        value=c("print(\\"Hello World!\\")",
                "print(\\"Goodbye cruel world...\\")"))
}
\seealso{
\code{vessel}, \code{moduleSource}, \code{source}
}

