% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{sessionID}
\alias{sessionID}
\title{Probably unique conduit session ID}
\format{\preformatted{ chr "conduit616a461dadd8"
}}
\usage{
sessionID
}
\description{
Probably unique conduit session ID
}
\seealso{
\code{parseModuleHost}
}
\keyword{datasets}

