% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/script.R
\name{sourceOrder}
\alias{sourceOrder}
\title{Determines running order for \code{moduleSource}s.}
\usage{
sourceOrder(sources)
}
\arguments{
\item{sources}{List of \code{moduleSource}s}
}
\value{
Running order as numeric vector
}
\description{
Determines running order for \code{moduleSource}s.
}
\details{
Order goes negative < 0 < no order given < positive.
}
\seealso{
\code{moduleSource}
}

