% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{startComponent}
\alias{startComponent}
\title{Return \code{pipe} start component}
\usage{
startComponent(pipe)
}
\arguments{
\item{pipe}{\code{pipe} object}
}
\value{
name of start component as string
}
\description{
Return \code{pipe} start component
}

