% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{startOutput}
\alias{startOutput}
\title{Return \code{pipe} start output}
\usage{
startOutput(pipe)
}
\arguments{
\item{pipe}{\code{pipe} object}
}
\value{
name of start output as string
}
\description{
Return \code{pipe} start output
}

