% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{topologicalSort}
\alias{topologicalSort}
\title{Do a topological sort on a graph}
\usage{
topologicalSort(edges)
}
\arguments{
\item{edges}{named list of edges from each vertex as character
vector}
}
\value{
character vector of sorted vertices
}
\description{
Return a linear ordering of vertices in an acyclic directed graph
such that for every directed edge from u to v, u comes before v in
the ordering.
}
\details{
\code{edges} list should be named for each vertex in the
graph. Each vertex object should be a character vector of vectors
to which the vector has a directed edge.

Uses Kahn's algorithm
\url{https://en.wikipedia.org/wiki/Topological_sorting#Kahn.27s_algorithm}
}

