% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validatePipe}
\alias{validatePipe}
\title{Validate a pipe}
\usage{
validatePipe(pipe, componentList)
}
\arguments{
\item{pipe}{\code{pipe} object}

\item{components}{\code{pipeline} components}
}
\value{
Logical value, TRUE if valid
}
\description{
Validates a single \code{pipe} against a \code{pipeline}'s
\code{component}s.
}
\details{
A pipe is valid if its startComponent, startOutput, endComponent
and endInput can all be matched by name to elements in the
\code{componentList}. The function should return a warning for each
invalid element.
}

