% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validatePipeline}
\alias{validatePipeline}
\title{Validate all pipes in a pipeline}
\usage{
validatePipeline(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} object}
}
\value{
Logical value, TRUE if valid
}
\description{
Validate all pipes in a pipeline
}

