% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vessel.R
\name{vessel}
\alias{vessel}
\title{Functions to manipulate vessels}
\description{
Functions to specify vessels for \code{moduleInput},
\code{moduleOutput}, or \code{moduleSource} objects.
}
\seealso{
\code{moduleInput}, \code{moduleOutput},
\code{moduleSource}
}

