% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{vesselToXML}
\alias{vesselToXML}
\title{Create XML corresponding to a \code{vessel} object.}
\usage{
vesselToXML(vessel, namespaceDefinitions = NULL, parent = NULL,
  addFinalizer = is.null(parent))
}
\arguments{
\item{vessel}{\code{vessel} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{parent}{parent XML object}

\item{addFinalizer}{logical add finalizer to free internal xmlDoc}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to a \code{vessel} object.
}
\seealso{
\code{vessel} objects, \code{XML::newXMLNode}
}

