% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{warnLanguageVersion}
\alias{warnLanguageVersion}
\title{Give warning if module execution violated moduleLanguage versions.}
\usage{
warnLanguageVersion(module, moduleResult)
}
\arguments{
\item{module}{\code{module} object}

\item{moduleResult}{\code{moduleResult} object}
}
\description{
If the version of the language used to execute a module's source
scripts has violated the \code{moduleLanguage}'s minVersion,
maxVersion or version, this function will raise a warning.
}
\seealso{
\code{runModule}
}

