% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addComponent}
\alias{addComponent}
\title{Add a new component to a pipeline}
\usage{
addComponent(newComponent, pipeline)
}
\arguments{
\item{newComponent}{\code{pipeline} or \code{module} object to be added}

\item{pipeline}{\code{pipeline} to be amended}
}
\value{
\code{pipeline} object
}
\description{
Add a new component to a pipeline
}

