% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{addPipe}
\alias{addPipe}
\title{Add a pipe to a pipeline}
\usage{
addPipe(newPipe, pipeline)
}
\arguments{
\item{newPipe}{\code{pipe} object}

\item{pipeline}{\code{pipeline} object}
}
\value{
\code{pipeline} object
}
\description{
This functions adds a new \code{pipe} to a \code{pipeline}.
}
\examples{
## create a pipeline with two modules
mod1 <- module(name = "setX", language = "R",
               description = "sets the value of x",
               outputs = list(
                   moduleOutput(
                       name = "x",
                       vessel = internalVessel("x"),
                       format = ioFormat("R character string"))),
               sources = list(
                   moduleSource(
                       vessel = scriptVessel("x <- \\"set\\""))))
mod2 <- module("showY", language = "R",
               description = "displays the value of Y",
               inputs = list(
                   moduleInput(
                       name = "y",
                       vessel = internalVessel("y"),
                       format = ioFormat("R character string"))),
               sources = list(
                   moduleSource(
                       vessel = scriptVessel("print(y)"))))
pline1 <- pipeline(name = "trivialpipeline", modules = list(mod1, mod2))
## create a pipe
pipe1 <- pipe("setX", "x",
              "showY", "y")
## add pipe to pipeline
pline1 <- addPipe(pipe1, pline1)
}
\seealso{
\code{pipe}, \code{pipeline}
}

