% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/component.R
\name{calculateOutputs.module}
\alias{calculateOutputs.module}
\title{calculate output objects produced by a module}
\usage{
\method{calculateOutputs}{module}(module, outputDirectory)
}
\description{
calculate output objects produced by a module
}

