% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{calculateOutputs}
\alias{calculateOutputs}
\title{Calculate ouput objects produced by a component}
\usage{
calculateOutputs(componentValue, outputDirectory)
}
\arguments{
\item{componentValue}{\code{module} or \code{pipeline} object}

\item{outputDirectory}{file location for component outputs}
}
\value{
named list of output objects
}
\description{
Calculate ouput objects produced by a component
}

