% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{calculateOutputs}
\alias{calculateOutputs}
\title{Calculate ouput objects produced by a pipeline \code{component}}
\usage{
calculateOutputs(component, outputDirectory)
}
\arguments{
\item{component}{\code{component} object}

\item{outputDirectory}{file location for component outputs}
}
\value{
named list of \code{output} objects
}
\description{
Calculate ouput objects produced by a pipeline \code{component}
}
\details{
As at 2016-01-19 a method for \code{pipeline} objects has
    not been implemented.
}

