% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{calculateOutputs.module}
\alias{calculateOutputs}
\alias{calculateOutputs.module}
\title{Calculate ouput objects produced by a \code{module} or
\code{pipeline} object}
\usage{
\method{calculateOutputs}{module}(object, outputDirectory)

calculateOutputs(object, outputDirectory)
}
\arguments{
\item{object}{\code{module} or \code{pipeline}}

\item{outputDirectory}{file location for component outputs}
}
\value{
named list of output objects
}
\description{
Calculate ouput objects produced by a \code{module} or
\code{pipeline} object
}
\details{
As at 2016-01-19 a method for \code{pipeline} objects has
    not been implemented.
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: calculate \code{output} objects produced by a \code{module}
}}

