% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python2.R, R/python3.R, R/script.R, R/shell.R
\name{command.RScript}
\alias{command}
\alias{command.RScript}
\alias{command.python2Script}
\alias{command.python3Script}
\alias{command.shellScript}
\title{Generate a system command to run a module's source scripts}
\usage{
\method{command}{RScript}(script)

\method{command}{python2Script}(script)

\method{command}{python3Script}(script)

command(script)

\method{command}{shellScript}(script)
}
\arguments{
\item{\code{script}}{object}
}
\value{
\code{command} list containing \code{command} and
    \code{args} character vectors
}
\description{
Generate a system command to run a module's source scripts
}
\details{
\code{script} should be the result of \code{prepareScript}

This function is usually called by \code{executeScript}.
}
\section{Methods (by class)}{
\itemize{
\item \code{RScript}: generate a command to execute an R language
script

\item \code{python2Script}: generate a command to execute a python2 language
script

\item \code{python3Script}: generate a command to execute a python3 language
script

\item \code{shellScript}: generate a command to execute a shell language
script
}}
\seealso{
\code{prepareScript}, \code{executeScript}
}

