% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/script.R
\name{command.RScript}
\alias{command}
\alias{command.RScript}
\title{Generate a system command to run a module's source scripts}
\usage{
\method{command}{RScript}(script)

command(script)
}
\arguments{
\item{\code{script}}{object}
}
\value{
\code{command} list containing \code{command} and
    \code{args} character vectors
}
\description{
Generate a system command to run a module's source scripts
}
\details{
\code{script} should be the result of \code{prepareScript}
}
\section{Methods (by class)}{
\itemize{
\item \code{RScript}: generate a command to execute an R language
script
}}

