% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/component.R
\name{component}
\alias{component}
\title{Create a component object}
\usage{
component(name, vessel = NULL, value)
}
\arguments{
\item{name}{Name of component}

\item{vessel}{\code{fileVessel} or \code{urlVessel}}

\item{value}{\code{pipeline} or \code{module} object}
}
\value{
\code{component} list containing:
  \item{name}{component name}
  \item{vessel}{source vessel}
  \item{value}{\code{pipeline} or \code{module} object}
}
\description{
Create a \code{component} object for use in a \code{pipeline}.
}
\details{
A component contains either a \code{pipeline} or
    \code{module} object in its \code{value} output. Optionally it
    can contain a \code{fileVessel} or \code{urlVessel} in its
    \code{vessel} element, referencing the resource from which the
    component was loaded.
}
\seealso{
\code{pipeline}, \code{module}
}

