% Generated by roxygen2 (4.0.1): do not edit by hand
\name{component}
\alias{component}
\title{Create a component object}
\usage{
component(name, type, value = NULL, ref = NULL, path = NULL)
}
\arguments{
\item{name}{Name of component}

\item{type}{Character value; \dQuote{pipeline} or \dQuote{module}}

\item{value}{\code{pipeline} or \code{module} object}

\item{ref}{xml filename}

\item{path}{location of xml file}
}
\value{
\code{component} list containing:
\item{name}{component name}
\item{ref}{xml filename}
\item{path}{path to xml file}
\item{type}{component type}
\item{value}{\code{pipeline} or \code{module} object}
}
\description{
Create a \code{component} object for use in a \code{pipeline}.
}
\details{
A \code{component} object is expected to provide either a \code{ref}, and
possibly a \code{path}, to an openapi XML fileIf an xml file, or a
\code{pipeline} or \code{module} object in \code{value}. Accordingly:

\itemize{
\item{if \code{ref} is given the resulting component will have
\code{value} coerced to NULL.}

\item{if no \code{ref} is given the resulting component will have \code{type}
coerced to the class of the \code{value} object.}
}

If \code{type} is not \dQuote{pipeline} or \dQuote{module} the function
will fail.
}

