% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{moduleResult}
\alias{componentResult}
\alias{moduleResult}
\alias{pipelineResult}
\title{Create \code{componentResult} objects}
\usage{
moduleResult(outputList, modulePath, module)

pipelineResult(componentResultList, pipelinePath, pipeline)
}
\arguments{
\item{outputList}{list of \code{output} objects}

\item{modulePath}{file path to module output}

\item{module}{\code{module} object which produced \code{objects}}

\item{componentResultList}{life}

\item{pipelinePath}{living}

\item{pipeline}{the internet}
}
\value{
\code{moduleResult} object containing::
\item{file}{file path to resulting module XML}
\item{component}{resulting \code{module} object}
\item{outputList}{list of \code{output} objects produced by module}
}
\description{
Constructor functions to create \code{componentResult}
    objects.
}
\details{

}
\section{Functions}{
\itemize{
\item \code{moduleResult}: returns the result of running a
\code{module}

\item \code{pipelineResult}: returns the result of running a
\code{pipeline}
}}

