% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{defaultIdfile}
\alias{defaultIdfile}
\title{Default idfile for ssh to remote hosts}
\format{\preformatted{ chr "/home/ashley/uoastorage/openapi/conduit/inst/conduit.key"
}}
\usage{
defaultIdfile
}
\description{
Default idfile for ssh to remote hosts
}
\keyword{datasets}

