% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{ensureModuleInput}
\alias{ensureModuleInput}
\title{convert module input to script}
\usage{
ensureModuleInput(input, resource, language)
}
\arguments{
\item{input}{\code{moduleInput} object}

\item{resource}{address of resource to be supplied as input}

\item{language}{language of module script}
}
\value{
character string containing script to ensure input or NULL
}
\description{
convert module input to script
}

