% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/python.R
\name{executeScript.python}
\alias{executeScript.python}
\title{Execute a script in the "python" language}
\usage{
\method{executeScript}{python}(script)
}
\description{
Execute a script in the "python" language
}

