% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{executeScript}
\alias{executeScript}
\title{Execute a prepared module script file.}
\usage{
executeScript(script, moduleHost = NULL, hostSubdir)
}
\arguments{
\item{script}{\code{script} object to be executed}

\item{moduleHost}{\code{moduleHost} object}

\item{hostSubdir}{output directory on \code{moduleHost}}
}
\value{
0 if successful
}
\description{
Execute a prepared module script file.
}
\details{
If \code{moduleHost} is provided script will be executed on
remote host in \code{hostSubdir} on that machine.

\code{hostSubdir} should be the result of running
\code{prepareModuleHost}
}
\seealso{
\code{moduleHost}, \code{prepareModuleHost}

\code{runModule}
}

