% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.R, R/python.R, R/script.R, R/shell.R
\name{executeScript.R}
\alias{executeScript}
\alias{executeScript.R}
\alias{executeScript.python}
\alias{executeScript.shell}
\title{Execute a prepared module script file.}
\usage{
\method{executeScript}{R}(script, host)

\method{executeScript}{python}(script)

executeScript(script, host)

\method{executeScript}{shell}(script)
}
\arguments{
\item{script}{script file to be executed}
}
\value{
named list of \code{moduleOutput} objects
}
\description{
Execute a prepared module script file.
}
\section{Methods (by class)}{
\itemize{
\item \code{R}: Execute a script in the "R" language

\item \code{python}: Execute a script in the "python" language

\item \code{shell}: Execute a script in the "shell" language
}}
\seealso{
\code{runModule}
}

