% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{executeScript}
\alias{executeScript}
\title{Execute a prepared module script file.}
\usage{
executeScript(script, host = NULL, hostSubdir)
}
\arguments{
\item{script}{script file to be executed}

\item{host}{list of host details}
}
\value{
0 if successful
}
\description{
Execute a prepared module script file.
}
\details{
If \code{host} is provided script will be executed on
remote host in \code{host$directory}.
}
\seealso{
\code{runModule}
}

