% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{executeScript}
\alias{executeScript}
\title{Execute a module source scripts.}
\usage{
executeScript(module, resources)
}
\arguments{
\item{module}{\code{module} object}

\item{resources}{Named list of input objects}
}
\value{
FIXME: nothing meaningful
}
\description{
Execute module source scripts in the language given by
\code{module$language}.
}
\details{
A module's inputs and outputs objects are calculated, and a
script is placed in the working directory. The function then attemps to
execute this script using the specified language.
}

