% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shell.R
\name{executeScript.shell}
\alias{executeScript.shell}
\title{Execute a script in the "shell" language}
\usage{
\method{executeScript}{shell}(script)
}
\description{
Execute a script in the "shell" language
}

