% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{expandSearchPaths}
\alias{expandSearchPaths}
\title{Expand a search path to an absolute path}
\usage{
expandSearchPaths(searchPaths, location = getwd())
}
\arguments{
\item{searchPaths}{search paths}

\item{location}{file directory of invoking pipeline/module xml (optional)}
}
\value{
absolute path as character string
}
\description{
Expand a search path to an absolute path
}

