% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{expandSearchPaths}
\alias{expandSearchPaths}
\title{Expand a search path to an absolute path}
\usage{
expandSearchPaths(s)
}
\arguments{
\item{s}{search path}
}
\value{
absolute path as character string
}
\description{
Expand a search path to an absolute path
}

