% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{exportComponent}
\alias{exportComponent}
\title{Export a component to an XML file}
\usage{
exportComponent(component, targetDirectory = getwd())
}
\arguments{
\item{component}{\code{component} object}

\item{targetDirectory}{File path for pipeline output}
}
\value{
\code{component} object
}
\description{
Saves an XML of the component in the
\code{targetDirectory}. Returns a component object to be used by a
\code{pipeline} XML file also in \code{targetDirectory}.
}

