% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exportComponent}
\alias{exportComponent}
\title{Export a component to an XML file}
\usage{
exportComponent(component, targetDirectory = getwd(),
  filename = paste0(component$name, ".xml"))
}
\description{
Export a component to an XML file
}

