% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportComponent}
\alias{exportComponent}
\title{Export a component to an XML file}
\usage{
exportComponent(component, targetDirectory = getwd(),
  filename = paste0(component$name, ".xml"))
}
\arguments{
\item{component}{\code{component} object}

\item{targetDirectory}{File path for pipeline output}

\item{filename}{Name of resulting file}
}
\value{
Resulting file path
}
\description{
Export a component to an XML file
}

