% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportPipeline}
\alias{exportPipeline}
\title{Wrie a pipeline and its modules to disk}
\usage{
exportPipeline(pipeline, targetDirectory)
}
\arguments{
\item{pipeline}{A \code{pipeline} list}

\item{targetDirectory}{Output directory path}
}
\value{
A list of the XML file paths written
}
\description{
Write a \code{pipeline} and its \code{module}s to disk as openapi XML files
}
\details{
Creates a directory named for the \code{pipeline} in \code{targetDirectory},
then saves \code{pipeline} and \code{module} XML files in this directory.

As at 2014-08-12 the \code{pipeline} is always saved as \file{pipeline.xml}
no matter what the \code{pipeline} name.
}

