% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{extractModuleSource.scriptVessel}
\alias{extractModuleSource}
\alias{extractModuleSource.fileVessel}
\alias{extractModuleSource.scriptVessel}
\title{Extract a module's source script}
\usage{
\method{extractModuleSource}{scriptVessel}(moduleSource)

\method{extractModuleSource}{fileVessel}(moduleSource)

extractModuleSource(moduleSource)
}
\arguments{
\item{vessel}{Vessel indicating source script location}
}
\value{
character vector of source scrip
}
\description{
Extract a module's source script
}
\section{Methods (by class)}{
\itemize{
\item \code{scriptVessel}: Extract a module's source script
from a scriptVessel

\item \code{fileVessel}: Extract a module's source script
from a fileVessel
}}
\seealso{
\code{executeScript}
}

