% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fetchRef}
\alias{fetchRef}
\title{Fetch the contents of a referenced file}
\usage{
fetchRef(ref, path)
}
\arguments{
\item{ref}{Address of referenced file}

\item{path}{File paths to search}
}
\value{
Character vector containing the contents of \code{ref}
}
\description{
Fetch the contents of a referenced file
}

