% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{fetchRef}
\alias{fetchRef}
\title{Read the contents of a referenced file}
\usage{
fetchRef(ref)
}
\arguments{
\item{ref}{character vector containing resolved ref location}
}
\value{
Character vector containing the contents of resource at
\code{ref}
}
\description{
Read the contents of a referenced file
}
\details{
\code{file} should be the result of the function
\code{resolveRef}. The class of this object determines which read
method is used.
}
\seealso{
\code{resolveRef}
}

