% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vessel.R
\name{fileVessel}
\alias{fileVessel}
\title{Create a \code{fileVessel} object.}
\usage{
fileVessel(ref, path = NULL)
}
\arguments{
\item{ref}{Absolute or relative location of file}

\item{path}{Optional search path for locating file}
}
\value{
\code{fileVessel}, \code{vessel} list object
}
\description{
Create a \code{fileVessel} object.
}
\details{
\code{fileVessel} objects are used to point to files on
the local filesystem.

These can be used to specify:

\itemize{
  \item{where a module script has created a file, as described in
        a \code{moduleOutput}}
  \item{from where a module script expects a file to be available,
        as described in a \code{moduleInput}}
  \item{from where a \code{moduleSource} is to be retrieved}
}

\code{ref} (and \code{path}, if given), must be character vectors of
length 1.
}
\examples{
data_file <- fileVessel(ref="~/myFiles/data.csv")
}
\seealso{
More about \code{vessel} objects, more about
\code{moduleInput}, \code{moduleOutput} and \code{moduleSource}
objects, an overview of \code{module} objects.
}

