% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/module.R, R/pipeline.R
\name{getDescription}
\alias{getDescription}
\alias{getDescription.module}
\alias{getDescription.pipeline}
\title{Extract object description}
\usage{
getDescription(x)

\method{getDescription}{module}(x)

\method{getDescription}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract object description
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Returns module description

\item \code{pipeline}: Returns pipeline description
}}

