% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/generic.R, R/module.R
\name{getLanguage}
\alias{getLanguage}
\alias{getLanguage.module}
\title{Retrieve object language}
\usage{
getLanguage(x)

\method{getLanguage}{module}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object language
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Returns module language
}}

