% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/module.R
\name{getLocation}
\alias{getLocation}
\alias{getLocation.module}
\title{Retrieve object location}
\usage{
getLocation(x)

\method{getLocation}{module}(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object location
}
\section{Methods (by class)}{
\itemize{
\item \code{module}: Returns location of module XML file
}}

