% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic.R, R/module.R, R/pipeline.R
\name{getName}
\alias{getName}
\alias{getName.module}
\alias{getName.pipeline}
\title{S3 generic to extract object name}
\usage{
getName(x)

\method{getName}{module}(x)

\method{getName}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
S3 generic to extract object name
}

