% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R, R/generic.R, R/module.R, R/pipeline.R
\name{getName.component}
\alias{getName}
\alias{getName.component}
\alias{getName.default}
\alias{getName.module}
\alias{getName.pipeline}
\title{Extract object name}
\usage{
\method{getName}{component}(x)

getName(x)

\method{getName}{default}(x)

\method{getName}{module}(x)

\method{getName}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract object name
}
\section{Methods (by class)}{
\itemize{
\item \code{component}: Returns component name

\item \code{default}: 

\item \code{module}: Returns module name

\item \code{pipeline}: Returns pipeline name
}}

