% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R, R/generic.R, R/module.R, R/pipeline.R
\name{getName.component}
\alias{getName}
\alias{getName.component}
\alias{getName.module}
\alias{getName.pipeline}
\title{Extract object name}
\usage{
getName.component(x)

getName(x)

getName(x)

getName.module(x)

getName.pipeline(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract object name
}
\section{Functions}{
\itemize{
\item \code{getName.component}: Returns component name

\item \code{getName}: 

\item \code{getName.module}: Returns module name

\item \code{getName.pipeline}: Returns pipeline name
}}

