% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/pipeline.R
\name{getPipes}
\alias{getPipes}
\alias{getPipes.pipeline}
\title{Extract pipes from object}
\usage{
getPipes(x)

\method{getPipes}{pipeline}(x)
}
\arguments{
\item{x}{object}
}
\description{
Extract pipes from object
}
\section{Methods (by class)}{
\itemize{
\item \code{pipeline}: Returns list of \code{pipe} objects
}}

