% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/component.R, R/generic.R
\name{getType.component}
\alias{getType}
\alias{getType.component}
\title{Retrieve object type}
\usage{
\method{getType}{component}(x)

getType(x)
}
\arguments{
\item{x}{object}
}
\description{
Retrieve object type
}
\section{Methods (by class)}{
\itemize{
\item \code{component}: Returns component type
}}

