% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getXMLAttr}
\alias{getXMLAttr}
\title{Gets a named XML attribute if it exists}
\usage{
getXMLAttr(xmlNode, attrName)
}
\arguments{
\item{xmlNode}{\code{XMLNode} whose attributes are to be extracted}

\item{attrName}{String vale naming attribute required}
}
\value{
Character string containing attribute value, or NULL
}
\description{
Gets a named XML attribute if it exists
}

