% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{graphPipeline}
\alias{graphPipeline}
\title{Create a \code{graphNEL} node-and-edge graph of a pipeline}
\usage{
graphPipeline(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} list object}
}
\value{
A \pkg{graph} \code{graphNEL} object
}
\description{
\code{graphPipeline} produces a directed graph of the given
\code{pipeline} with components as nodes and pipes as directed edges.
}
\details{
The edges in the resulting graph object are a dumb representation
of the pipes in a pipeline, i.e. they do not retain output and
input information. They merely serve to provide on object for a
topological sort for the purposes of determining the execution
order of a pipeline's components.
}

