% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graphPipeline}
\alias{graphPipeline}
\title{Create a graph of a pipeline}
\usage{
graphPipeline(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} list object}
}
\value{
A \pkg{graph} \code{graphNEL} object
}
\description{
\code{graphPipeline} produces a directed graph of the given
\code{pipeline} with components as nodes and pipes as directed edges
}

