% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{inputObjects}
\alias{inputObjects}
\title{Returns a named list of input objects}
\usage{
inputObjects(pipes, components, pipelinePath)
}
\arguments{
\item{pipes}{List of \code{pipe}s}

\item{components}{List of \code{component}s}

\item{pipelinePath}{Absolute file path to originating \code{pipeline}
XML file}
}
\value{
named list of file locations by input names
}
\description{
\code{inputsList} returns a named list of absolute file locations for
components' inputs.
}
\details{
List items are named as COMPONENT_NAME.INPUT_NAME
}

