% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inputsList}
\alias{inputsList}
\title{Returns a named list of input addresses}
\usage{
inputsList(pipes, components, pipelinePath)
}
\arguments{
\item{pipes}{List of \code{pipe}s}

\item{components}{List of \code{component}s}

\item{pipelinePath}{Absolute file path to currently running \code{pipeline}}
}
\description{
\code{inputsList} returns a named list of absolute file locations for
components' inputs.
}
\details{
List items are named as COMPONENT_NAME.INPUT_NAME
}

