% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{internalExtension}
\alias{internalExtension}
\title{Returns the correct file extension for a language's 'internal' files}
\usage{
internalExtension(language)
}
\arguments{
\item{language}{language name}
}
\value{
file exension as character as ".EXT"
}
\description{
Returns the correct file extension for a language's 'internal' files
}

