% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{internalExtension}
\alias{internalExtension}
\title{Returns the correct file extension for a language's 'internal' files}
\usage{
internalExtension(moduleLanguage)
}
\arguments{
\item{moduleLanguage}{\code{moduleLanguage} object}
}
\value{
file exension as character as ".EXT"
}
\description{
Returns the correct file extension for a language's 'internal' files
}

