% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{internalExtension}
\alias{internalExtension}
\title{Returns the correct file extension for a platform's 'internal' files}
\usage{
internalExtension(language)
}
\arguments{
\item{language}{language name}
}
\value{
files exension as character as ".EXT"
}
\description{
Returns the correct file extension for a platform's 'internal' files
}

