% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python.R, R/script.R, R/shell.R
\name{internalInputScript.R}
\alias{internalInputScript}
\alias{internalInputScript.R}
\alias{internalInputScript.python}
\alias{internalInputScript.shell}
\title{prepare script to resolve internal input}
\usage{
internalInputScript.R(symbol)

internalInputScript.python(symbol)

internalInputScript.shell(symbol)
}
\arguments{
\item{symbol}{character string with class set to language of module script}
}
\value{
character vector of script to ensure input
}
\description{
prepare script to resolve internal input
}
\section{Functions}{
\itemize{
\item \code{internalInputScript.R}: prepare internal input script for R
language

\item \code{internalInputScript.python}: prepare internal input script for
python language

\item \code{internalInputScript.shell}: prepare internal input script for
shell language
}}

