% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python.R, R/script.R, R/shell.R
\name{internalInputScript.R}
\alias{internalInputScript}
\alias{internalInputScript.R}
\alias{internalInputScript.python}
\alias{internalInputScript.shell}
\title{prepare script to resolve internal input}
\usage{
\method{internalInputScript}{R}(symbol, inputObject)

\method{internalInputScript}{python}(symbol, inputObject)

internalInputScript(symbol, inputObject)

\method{internalInputScript}{shell}(symbol, inputObject)
}
\arguments{
\item{symbol}{character string with class set to language of module script}

\item{inputObject}{file location of serialised language object}
}
\value{
character vector of script to ensure input
}
\description{
prepare script to resolve internal input
}
\section{Methods (by class)}{
\itemize{
\item \code{R}: prepare internal input script for R
language

\item \code{python}: prepare internal input script for
python language

\item \code{shell}: prepare internal input script for
shell language
}}

