% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalInputScript}
\alias{internalInputScript}
\title{prepare internal input script}
\usage{
internalInputScript(symbol, inputObject)
}
\arguments{
\item{symbol}{character string with class set to language of module script}

\item{inputObject}{file location of serialised language object}
}
\value{
character vector of script to ensure input
}
\description{
prepare internal input script
}

