% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python.R, R/script.R, R/shell.R
\name{internalInputScript.RSymbol}
\alias{internalInputScript}
\alias{internalInputScript.RSymbol}
\alias{internalInputScript.pythonSymbol}
\alias{internalInputScript.shellSymbol}
\title{Prepare script for internal inputs}
\usage{
\method{internalInputScript}{RSymbol}(symbol)

\method{internalInputScript}{pythonSymbol}(symbol)

internalInputScript(symbol)

\method{internalInputScript}{shellSymbol}(symbol)
}
\arguments{
\item{symbol}{\code{symbol} object}
}
\value{
character vector of script to ensure input
}
\description{
These functions prepare a module script snippet to resolve an
internal input
}
\section{Methods (by class)}{
\itemize{
\item \code{RSymbol}: prepare internal input script for R
language

\item \code{pythonSymbol}: prepare internal input script for
python language

\item \code{shellSymbol}: prepare internal input script for
shell language
}}

