% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python2.R, R/python3.R, R/script.R, R/shell.R
\name{internalInputScript.RSymbol}
\alias{internalInputScript}
\alias{internalInputScript.RSymbol}
\alias{internalInputScript.python2Symbol}
\alias{internalInputScript.python3Symbol}
\alias{internalInputScript.shellSymbol}
\title{Prepare script for internal inputs}
\usage{
\method{internalInputScript}{RSymbol}(symbol)

\method{internalInputScript}{python2Symbol}(symbol)

\method{internalInputScript}{python3Symbol}(symbol)

internalInputScript(symbol)

\method{internalInputScript}{shellSymbol}(symbol)
}
\arguments{
\item{symbol}{\code{symbol} object}
}
\value{
script as character vector
}
\description{
These functions prepare a module script snippet to resolve an
internal input
}
\section{Methods (by class)}{
\itemize{
\item \code{RSymbol}: prepare internal input script for R
language

\item \code{python2Symbol}: prepare internal input script for
python2 language

\item \code{python3Symbol}: prepare internal input script for
python3 language

\item \code{shellSymbol}: prepare internal input script for
shell language
}}

