% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{internalOutputScript}
\alias{internalOutputScript}
\title{prepare script to resolve internal output}
\usage{
internalOutputScript(symbol)
}
\arguments{
\item{symbol}{character string with class set to language of module script}
}
\value{
character vector of script to ensure input
}
\description{
prepare script to resolve internal output
}

