% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/R.R, R/python.R, R/script.R, R/shell.R
\name{internalOutputScript.R}
\alias{internalOutputScript}
\alias{internalOutputScript.R}
\alias{internalOutputScript.python}
\alias{internalOutputScript.shell}
\title{prepare script to resolve internal output}
\usage{
\method{internalOutputScript}{R}(symbol)

\method{internalOutputScript}{python}(symbol)

internalOutputScript(symbol)

\method{internalOutputScript}{shell}(symbol)
}
\arguments{
\item{symbol}{character string with class set to language of module script}
}
\value{
character vector of script to ensure input
}
\description{
prepare script to resolve internal output
}
\section{Methods (by class)}{
\itemize{
\item \code{R}: create script to create internal
output for language = "R"

\item \code{python}: create script to create internal
output for language = "python"

\item \code{shell}: create script to create internal
output for language = "shell"
}}

