% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R, R/python2.R, R/python3.R, R/script.R, R/shell.R
\name{internalOutputScript.RSymbol}
\alias{internalOutputScript}
\alias{internalOutputScript.RSymbol}
\alias{internalOutputScript.python2Symbol}
\alias{internalOutputScript.python3Symbol}
\alias{internalOutputScript.shellSymbol}
\title{prepare script to resolve internal output}
\usage{
\method{internalOutputScript}{RSymbol}(symbol)

\method{internalOutputScript}{python2Symbol}(symbol)

\method{internalOutputScript}{python3Symbol}(symbol)

internalOutputScript(symbol)

\method{internalOutputScript}{shellSymbol}(symbol)
}
\arguments{
\item{symbol}{\code{symbol} object}
}
\value{
script as character vector
}
\description{
These functions prepare a module script snippet to produce an
internal output
}
\section{Methods (by class)}{
\itemize{
\item \code{RSymbol}: create script to create internal
output for language = "R"

\item \code{python2Symbol}: create script to create internal
output for language = "python2"

\item \code{python3Symbol}: create script to create internal
output for language = "python3"

\item \code{shellSymbol}: create script to create internal
output for language = "shell"
}}

