% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{ioFormatToXML}
\alias{ioFormatToXML}
\title{Create XML corresponding to an \code{ioFormat} object}
\usage{
ioFormatToXML(ioFormat, namespaceDefinitions = NULL)
}
\arguments{
\item{ioFormat}{\code{ioFormat} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to an \code{ioFormat} object
}
\seealso{
\code{ioFormat} objects
}

