% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/component.R
\name{loadComponent}
\alias{loadComponent}
\title{Load a component's value from XML}
\usage{
loadComponent(component)
}
\arguments{
\item{component}{A \code{component} object}
}
\value{
\code{component} object with the corresponding \code{pipeline}
or \code{module} object its value slot
}
\description{
Load a \code{component}'s value from the XML file given in its ref and
path slots.
}

