% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadModule}
\alias{loadModule}
\title{Read a module from an XML file}
\usage{
loadModule(name, ref, path = searchPaths, namespaces = c(oa =
  "http://www.openapi.org/2014/"))
}
\arguments{
\item{name}{Name of module}

\item{ref}{File name of module XML document}

\item{path}{File path to module XML document}

\item{namespaces}{Namespaces used in XML document as named character vector}
}
\value{
\code{module} list
}
\description{
Read a module from an XML file
}

