% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadPipeline}
\alias{loadPipeline}
\title{Load a pipeline and its modules from disk}
\usage{
loadPipeline(filename,
  pipelineName = tools::file_path_sans_ext(basename(filename)),
  namespaces = c(oa = "http://www.openapi.org/2014/"))
}
\arguments{
\item{filename}{file path to \code{pipeline} XML}

\item{pipelineName}{as character value}

\item{namespaces}{named charactor vector}
}
\value{
\code{pipeline} list containing:
\item{name}{as character value}
\item{description}{as character value}
\item{components}{list of \code{module}s and \code{pipeline}s}
\item{pipes}{list of \code{pipe}s}
}
\description{
Read an openapi \code{pipeline} and its associated \code{module}s from an
XML document
}

