% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{loadPipeline}
\alias{loadPipeline}
\title{Load a pipeline from an XML file}
\usage{
loadPipeline(name, ref, path = NULL, namespaces = c(oa =
  "http://www.openapi.org/2014/"))
}
\arguments{
\item{name}{Name of pipeline}

\item{ref}{Path to XML file}

\item{path}{Search path (optional)}

\item{namespaces}{Namespaces used in XML document}
}
\value{
\code{pipeline} list
}
\description{
Reads an XML file given by \code{ref} and \code{path} and interprets to
produce a \code{pipeline}.
}
\details{
If \code{path} is not set and conduit needs to search for the file the
default search paths are used.
}
\examples{
pln1xml <- system.file("extdata", "simpleGraph", "simpleGraph-pipeline.xml",
                       package = "conduit")
pln1 <- loadPipeline(name = "simpleGraph", ref = pln1xml)
}
\seealso{
\code{pipeline}
}

