% Generated by roxygen2 (4.0.2): do not edit by hand
\name{module}
\alias{module}
\title{Create a \code{module} list object}
\usage{
module(name, description = "", platform, inputs = list(),
  outputs = list(), sources = list(), ref = NULL, path = NULL)
}
\arguments{
\item{name}{Name of module}

\item{description}{A basic description of the module}

\item{platform}{Platform name}

\item{inputs}{List of \code{moduleInput} objects}

\item{outputs}{List of \code{moduleOutput} objects}

\item{sources}{List of \code{moduleSource} objects}

\item{ref}{URI or filename of module XML file (optional)}

\item{path}{Search path for module XML file (optional)}
}
\value{
\code{module}
}
\description{
Create a \code{module} list object
}

