% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{moduleHostToXML}
\alias{moduleHostToXML}
\title{Create XML corresponding to a \code{moduleHost} object}
\usage{
moduleHostToXML(moduleHost)
}
\description{
Create XML corresponding to a \code{moduleHost} object
}

