% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerHost.R, R/module.R, R/moduleInputHost.R, R/vagrantHost.R
\name{moduleHostToXML.dockerHost}
\alias{moduleHostToXML}
\alias{moduleHostToXML.dockerHost}
\alias{moduleHostToXML.moduleInputHost}
\alias{moduleHostToXML.vagrantHost}
\title{Create XML corresponding to a \code{moduleHost} object}
\usage{
\method{moduleHostToXML}{dockerHost}(moduleHost)

moduleHostToXML(moduleHost)

\method{moduleHostToXML}{moduleInputHost}(moduleInputHost)

\method{moduleHostToXML}{vagrantHost}(moduleHost)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}
}
\value{
XML node representing module host
}
\description{
Create XML corresponding to a \code{moduleHost} object
}
\section{Methods (by class)}{
\itemize{
\item \code{dockerHost}: Create XML corresponding to a
\code{dockerHost} object

\item \code{moduleInputHost}: Create XML corresponding to a
\code{moduleInputHost} object

\item \code{vagrantHost}: Create XML corresponding to a
\code{vagrantHost} object
}}

