% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleInput}
\alias{moduleInput}
\title{Create a \code{module} input}
\usage{
moduleInput(name, type, format = "", formatType = "text")
}
\arguments{
\item{name}{Input name}

\item{type}{\sQuote{internal} or \sQuote{external}}

\item{format}{Input format}

\item{formatType}{Defaults to \dQuote{text}}
}
\value{
named \code{moduleInput} character vector containing:
\itemize{
  \item{name}
  \item{type}
  \item{format}
  \item{formatType}
}
}
\description{
Creates a \code{moduleInput} vector for use in a \code{module}'s inputs list.
}
\details{
\code{type} must be \sQuote{internal} or \sQuote{external}.
}
\examples{
inp1 <- moduleInput(name = "bigData", type = "internal",
                    format = "R data frame")
}
\seealso{
\code{module}
}

