% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleInput}
\alias{moduleInput}
\title{Create a \code{module} input}
\usage{
moduleInput(name, vessel, format)
}
\arguments{
\item{name}{Input name}

\item{vessel}{\code{vessel} object}

\item{format}{\code{ioFormat} object}
}
\value{
named \code{moduleInput} list containing:
\itemize{
  \item{name}
  \item{vessel}
  \item{format}
}
}
\description{
Create a \code{moduleInput} list for use in a \code{module}'s inputs list
}
\examples{
internalInput <-
    moduleInput(
        name = "bigData",
        vessel = internalVessel(symbol = "bigData"),
        format = ioFormat("R data frame"))
fileInput <-
    moduleInput(
        name = "scores.csv",
        vessel = fileVessel(ref = "2015-03-13-scores.csv"),
        format = ioFormat("CSV file"))
}
\seealso{
\code{module}
}

