% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleInput}
\alias{moduleInput}
\title{Create a \code{module} input node}
\usage{
moduleInput(name, type, format = "", formatType = "text")
}
\arguments{
\item{name}{Input name}

\item{type}{\sQuote{internal} or \sQuote{external}}

\item{format}{Input format}

\item{formatType}{Defaults to \dQuote{text}}
}
\description{
Create a \code{module} input node
}
\seealso{
\code{module}
}

