% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleIOToXML}
\alias{moduleIOToXML}
\title{Create XML corresponding to \code{moduleIO} object}
\usage{
moduleIOToXML(moduleIO, namespaceDefinitions = NULL)
}
\arguments{
\item{moduleIO}{\code{moduleIO} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to \code{moduleIO} object
}
\seealso{
\code{moduleIO} objects
}

