% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleOutput}
\alias{moduleOutput}
\title{Create a \code{module} output input}
\usage{
moduleOutput(name, type, format = "", formatType = "text", ref = "")
}
\arguments{
\item{name}{Output name}

\item{type}{\sQuote{internal} or \sQuote{external}}

\item{format}{Output format}

\item{formatType}{Defaults to \dQuote{text}}

\item{ref}{Filename of \sQuote{external} output}
}
\description{
Creates a \code{moduleOutput} vector for use in a \code{module}'s outputs
list.
}
\details{
\code{type} must be \sQuote{internal} or \sQuote{external}.

It \code{type} is \dQuote{external}, a \code{ref} is required. This needs
to be a resolveable URI created by the \code{module}'s source(s).
}
\seealso{
\code{module}
}

