% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleOutput}
\alias{moduleOutput}
\title{Create a \code{module} output}
\usage{
moduleOutput(name, vessel, format)
}
\arguments{
\item{name}{Output name}

\item{vessel}{\code{vessel} object}

\item{format}{\code{ioFormat} object}
}
\value{
named \code{moduleOutput} list containing:
\itemize{
  \item{name}
  \item{vessel}
  \item{format}
}
}
\description{
Create a \code{moduleOutput} list for use in a \code{module}'s
outputs list.
}
\details{
\code{vessel} cannot be a \code{scriptVessel} object, as these are not
defined for \code{moduleIO} objects.
}
\examples{
internalOutput <-
    moduleOutput(
        name = "bigData",
        vessel = internalVessel(symbol = "bigData"),
        format = ioFormat("R data frame"))
fileInput <-
    moduleOutput(
        name = "scores.csv",
        vessel = fileVessel(ref = "2015-03-13-scores.csv"),
        format = ioFormat("CSV file"))
}
\seealso{
\code{module}
}

