% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleOutput}
\alias{moduleOutput}
\title{Create a \code{module} output node}
\usage{
moduleOutput(name, type, format = "", formatType = "text", ref = "")
}
\arguments{
\item{name}{Output name}

\item{type}{\sQuote{internal} or \sQuote{external}}

\item{format}{Output format}

\item{formatType}{Defaults to \dQuote{text}}

\item{ref}{Filename of \sQuote{external} output}
}
\description{
Create a \code{module} output node
}
\seealso{
\code{module}
}

