% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modulePlatform}
\alias{modulePlatform}
\title{Create a \code{module} platform node}
\usage{
modulePlatform(name)
}
\arguments{
\item{name}{Name of platform}
}
\value{
A named character vector containing the platform name
}
\description{
Create a \code{module} platform node
}

