% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{moduleResult}
\alias{moduleResult}
\title{Produce a \code{moduleResult} object to be returned by
\code{runModule}.}
\usage{
moduleResult(objects, modulePath, module)
}
\arguments{
\item{objects}{list of \code{output} objects}

\item{modulePath}{file path to module output}

\item{module}{\code{module} object which produced \code{objects}}
}
\value{
\code{moduleResult}
}
\description{
Produce a \code{moduleResult} object to be returned by
\code{runModule}.
}

