% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moduleSource}
\alias{moduleSource}
\title{Create a \code{module} source node}
\usage{
moduleSource(value, ref = NULL, path = defaultSearchPaths, type = "",
  order = "")
}
\arguments{
\item{value}{}

\item{ref}{}

\item{path}{}

\item{type}{}

\item{order}{}
}
\description{
Create a \code{module} source node
}
\seealso{
\code{module}
}

