% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleSource}
\alias{moduleSource}
\title{Create a \code{module} source}
\usage{
moduleSource(vessel, order = NULL)
}
\arguments{
\item{vessel}{\code{vessel} object}

\item{order}{numeric value specifying source position in sources}
}
\value{
named \code{moduleSource} list containing:
\itemize{
  \item{vessel: \code{vessel} object}
  \item{order: numeric value determining position of source in sources}
}
}
\description{
Create a \code{moduleSource} object for use in a \code{module}'s sources
list.
}
\details{
The contents of a module's source script are provided by
the \code{vessel} object. For inline code use a \code{scriptVessel}
object. For a script file on the local filesystem use
\code{fileVessel}.

\code{module} sources are exectuted in the order determined by each
source's \sQuote{order}.

Running order is:
\enumerate{
  \item{negative numbers in ascending order}
  \item{zero}
  \item{no order specified}
  \item{positive numbers in ascending order}
}
}
\examples{
## create moduleSource with source script inline
val1 <- c("x <- 1:10", "y <- rnorm(10, 0, 1)", "plot(x, y)")
src1 <- moduleSource(vessel = scriptVessel(value = val1),
                     order = -1)

## create a moduleSource with source script from file
modScript <- system.file("extdata", "simpleGraphScripts", "createGraph.R",
                         package = "conduit")
src2 <- moduleSource(vessel = fileVessel(ref = modScript))
}
\seealso{
\code{module}, \code{fileVessel}, \code{scriptVessel}
}

