% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleSourceToXML}
\alias{moduleSourceToXML}
\title{Create XML corresponding to \code{moduleSource} object}
\usage{
moduleSourceToXML(moduleSource, namespaceDefinitions = NULL)
}
\arguments{
\item{moduleSource}{\code{moduleSource} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLInternalNode} object
}
\description{
Create XML corresponding to \code{moduleSource} object
}
\seealso{
\code{moduleSource} objects
}

