% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleToXML}
\alias{moduleToXML}
\title{Convert a \code{module} object to XML}
\usage{
moduleToXML(module, namespaceDefinitions = NULL)
}
\arguments{
\item{module}{\code{module} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLInternalNode} object
}
\description{
Convert a \code{module} object to XML
}
\seealso{
\code{module} objects
}

