% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{moduleToXML}
\alias{moduleToXML}
\title{Convert a module to XML}
\usage{
moduleToXML(module, namespaceDefinitions = NULL)
}
\arguments{
\item{module}{\code{module} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLNode} object
}
\description{
Convert a module to XML
}

