% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{output}
\alias{output}
\title{return \code{output} produced by a \code{moduleOutput}}
\usage{
output(moduleOutput, language, outputDirectory)
}
\arguments{
\item{moduleOutput}{\code{moduleOutput} object}

\item{language}{module script language}

\item{outputDirectory}{file location for module execution}
}
\value{
\code{output} list object, containing:

\item{name}{output name}
\item{format}{\code{ioFormat} object}
\item{vessel}{\code{vessel} object}
\item{language}{module language}
\item{ref}{address of output object produced}
}
\description{
Return the object produced by a module output.
}
\details{
This function returns a reference to the object produced by a
module's output when the module is executed. The reference
contained in this object is not guaranteed to exist until after
module execution.
}

