% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{outputObject}
\alias{outputObject}
\title{return output object produced by a module output}
\usage{
outputObject(output, language, outputDirectory)
}
\arguments{
\item{output}{\code{moduleOutput} object}

\item{language}{module script language}

\item{outputDirectory}{file location for module execution}
}
\value{
output object
}
\description{
return output object produced by a module output
}

