% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{parseModuleHost}
\alias{parseModuleHost}
\title{Parse a module's host}
\usage{
parseModuleHost(host, idfile = get("defaultIdfile", envir = .conduit.global))
}
\arguments{
\item{host}{module host as character string}

\item{idfile}{authentication key}
}
\value{
list of \itemize{
  \item user
  \item password
  \item address
  \item port
  \item directory
  \item idfile}
}
\description{
Parse the the host attribute from a \code{module} object.
}
\details{
Default user = 'conduit', default port = '22'.

\code{directory} slot is filled by a random output directory to be
created on the remote host at \code{/tmp/{sessionID}/}
}

