% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pipe}
\alias{pipe}
\title{Create a pipe object}
\usage{
pipe(startComponentName = NULL, startOutputName, endComponentName = NULL,
  endInputName, startComponentRef = NULL, endComponentRef = NULL)
}
\arguments{
\item{startComponentName}{Name of start module}

\item{startOutputName}{Name of start object}

\item{endComponentName}{Name of end module}

\item{endInputName}{Name of end input}

\item{startComponentRef}{Address of start module}

\item{endComponentRed}{Address of end module}
}
\value{
\code{pipe} connecting \code{startComponentName}.\code{startOutputName} to \code{endComponentName}.\code{endInputName}
}
\description{
Create a pipe object
}

