% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pipe}
\alias{pipe}
\title{Creates a \code{pipe} object}
\usage{
pipe(startComponent, startOutput, endComponent, endInput)
}
\arguments{
\item{startComponent}{Name of start component}

\item{startOutput}{Name of start output}

\item{endComponent}{Name of end module}

\item{endInput}{Name of end input}
}
\value{
\code{pipe} connecting \code{startComponentName}.\code{startOutputName} to \code{endComponentName}.\code{endInputName}
}
\description{
Creates a \code{pipe} object which connects the \code{startComponent}'s
\code{startOutput} to the \code{endComponent}'s \code{endInput}.
}
\examples{
pipe1 <- pipe(startComponent = "setX", startOutput = "x",
              endComponent = "showY", endInput = "y")
}
\seealso{
\code{pipeline}, \code{addPipe}
}

