% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pipelineToXML}
\alias{pipelineToXML}
\title{Convert a pipeline to XML}
\usage{
pipelineToXML(pipeline, namespaceDefinitions = NULL, export = FALSE)
}
\arguments{
\item{pipeline}{\code{pipeline} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{export}{if true inline components are converted to reference files}
}
\value{
\code{XMLNode} object
}
\description{
Convert a pipeline to XML
}

