% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pipelineToXML}
\alias{pipelineToXML}
\title{Convert a pipeline to XML}
\usage{
pipelineToXML(pipeline, namespaceDefinitions = NULL, export = FALSE)
}
\arguments{
\item{pipeline}{\code{pipeline} object}

\item{namespaceDefinition}{XML namespaces as character vector}
}
\value{
\code{XMLNode} object
}
\description{
Convert a pipeline to XML
}

