% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipelineToXML}
\alias{pipelineToXML}
\title{Convert a pipeline to XML}
\usage{
pipelineToXML(pipeline, namespaceDefinitions = NULL)
}
\arguments{
\item{pipeline}{\code{pipeline} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{XMLNode} object
}
\description{
Convert a pipeline to XML
}

