% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipelineToXML}
\alias{pipelineToXML}
\title{Convert a pipeline to XML}
\usage{
pipelineToXML(pipeline, namespaceDefinitions = NULL, export = FALSE)
}
\arguments{
\item{pipeline}{\code{pipeline} object}

\item{namespaceDefinitions}{XML namespaces as character vector}

\item{export}{logical, indicating whether to export inlince components as
referenced XML files}
}
\value{
\code{XMLNode} object
}
\description{
Convert a pipeline to XML
}

