% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipeToXML}
\alias{pipeToXML}
\title{Convert a pipe to XML}
\usage{
pipeToXML(pipe, namespaceDefinitions = NULL)
}
\arguments{
\item{pipe}{\code{pipe} object}

\item{namespaceDefinitions}{XML namespaces as character vector}
}
\value{
\code{xmlNode} object
}
\description{
Convert a pipe to XML
}

