% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScript}
\alias{prepareScript}
\title{Prepare a script for executing a module in its language.}
\usage{
prepareScript(module)
}
\arguments{
\item{module}{\code{module} object}
}
\value{
\code{script} object naming script file
}
\description{
This function creates an executable script file from a
\code{module} object.
}
\details{
The script returned will include code to load internal inputs,
followed by the module source scripts in the correct order, and
ending with code to produce internal outputs for consumption by
other modules.

The resulting script is saved to the current working directory.
}

