% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScript}
\alias{prepareScript}
\title{Prepare a script for executing a module in its language.}
\usage{
prepareScript(module, inputObjects)
}
\arguments{
\item{module}{\code{module} object}

\item{inputObjects}{Named list of input objects}
}
\value{
List object containg scriptPath and host, with class set to
module$language.
}
\description{
Prepare a script for executing a module in its language.
}
\details{
Resolves the module's internal inputs and creates a script
file from the supplied \code{module},
}

