% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScriptInput}
\alias{prepareScriptInput}
\title{Prepare script to create inputs}
\usage{
prepareScriptInput(moduleInput, moduleLanguage)
}
\arguments{
\item{moduleInput}{module input object}

\item{moduleLanguage}{\code{moduleLanguage} object}
}
\value{
Script as character vector
}
\description{
Prepare script to create inputs
}
\details{
if a module input is to be fulfilled via an internalVessel
    the module source scripts will require the symbol to be loaded
    prior to execution. other vessel types do not need to be loaded
    in script.
}

