% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/script.R
\name{prepareScriptOutput}
\alias{prepareScriptOutput}
\title{Prepare script to create outputs}
\usage{
prepareScriptOutput(output, language)
}
\arguments{
\item{output}{output name}

\item{language}{module language}
}
\value{
Script as character vector
}
\description{
Prepare script to create outputs
}

