% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readComponentNode}
\alias{readComponentNode}
\title{Parse a component \code{xmlNode} and return a \code{component}.}
\usage{
readComponentNode(node, pipelinePath)
}
\arguments{
\item{node}{An \code{xmlNode} named \dQuote{component}.}
}
\value{
\code{component} object
}
\description{
Parse a component \code{xmlNode} and return a \code{component}.
}

