% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{readComponentNode}
\alias{readComponentNode}
\title{Parse a component \code{xmlNode} and return a \code{component}.}
\usage{
readComponentNode(node, location)
}
\arguments{
\item{node}{An \code{xmlNode} named \dQuote{component}.}

\item{location}{file directory of invoking pipeline/module xml (optional)}
}
\value{
\code{component} object
}
\description{
Parse a component \code{xmlNode} and return a \code{component}.
}

