% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleHostXML}
\alias{readModuleHostXML}
\title{Create a \code{moduleHost} object from host XML}
\usage{
readModuleHostXML(moduleHostXML)
}
\arguments{
\item{moduleHostXML}{host XML}
}
\value{
\code{moduleHost} object
}
\description{
Create a \code{moduleHost} object from host XML
}

