% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleIOXML}
\alias{readModuleIOXML}
\title{Create a \code{moduleIO} object from input/output XML}
\usage{
readModuleIOXML(xml)
}
\arguments{
\item{xml}{input/output XML}
}
\value{
\code{moduleIO} object
}
\description{
Create a \code{moduleIO} object from input/output XML
}

