% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readModuleXML}
\alias{readModuleXML}
\title{Parse module XML and return a module object}
\usage{
readModuleXML(name, xml, path = NULL)
}
\arguments{
\item{name}{Module name}

\item{xml}{Module \code{XMLNode}}
}
\value{
\code{module} object
}
\description{
Parse module XML and return a module object
}

