% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleXML}
\alias{readModuleXML}
\title{Parse module XML and return a \code{module} object}
\usage{
readModuleXML(name, xml, location = getwd())
}
\arguments{
\item{name}{module name}

\item{xml}{module \code{XMLNode}}

\item{location}{file directory of invoking pipeline/module xml (optional)}
}
\value{
\code{module} object
}
\description{
Parse module XML and return a \code{module} object
}
\details{
If <host/> contains <moduleInput/> function fails if
    moduleInput name does not match any input names.
}

