% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/module.R
\name{readModuleXML}
\alias{readModuleXML}
\title{Parse module XML and return a module object}
\usage{
readModuleXML(name, xml, location = getwd())
}
\arguments{
\item{name}{module name}

\item{xml}{module \code{XMLNode}}

\item{location}{file directory of invoking pipeline/module xml (optional)}
}
\value{
\code{module} object
}
\description{
Parse module XML and return a module object
}

