% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{readPipelineXML}
\alias{readPipelineXML}
\title{Parse pipeline XML and return a pipeline object}
\usage{
readPipelineXML(name, xml, location = getwd())
}
\arguments{
\item{name}{Pipeline name}

\item{xml}{Pipeline \code{XMLNode}}

\item{location}{file directory of invoking pipeline/module xml (optional)}
}
\value{
\code{pipeline} object
}
\description{
Parse pipeline XML and return a pipeline object
}

