% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readPipelineXML}
\alias{readPipelineXML}
\title{Parse pipeline XML and return a pipeline object}
\usage{
readPipelineXML(name, xml, path = NULL)
}
\arguments{
\item{name}{Pipeline name}

\item{xml}{Pipeline \code{XMLNode}}

\item{path}{Search path (optional)}
}
\value{
\code{pipeline} object
}
\description{
Parse pipeline XML and return a pipeline object
}

