% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{resolveOutput}
\alias{resolveOutput}
\title{Checks a module output object has been created.}
\usage{
resolveOutput(output, language, host, outputDirectory = getwd())
}
\arguments{
\item{output}{\code{moduleOutput} object}

\item{language}{module language}

\item{host}{host list}

\item{internalExtension}{file extension for serialized internal language
object}
}
\value{
named list containing:
\itemize{
  \item name: object name
  \item type: object vessel type
  \item object: output object
}
}
\description{
Checks a module output object has been created.
}
\details{
Will produce an error if the object does not exist.
}

