% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R, R/vagrantHost.R
\name{retrieveModuleHost}
\alias{retrieveModuleHost}
\alias{retrieveModuleHost.vagrantHost}
\title{Retrieve results of running a module on a remote host}
\usage{
retrieveModuleHost(moduleHost, outputLocation, modulePath)

\method{retrieveModuleHost}{vagrantHost}(moduleHost, outputLocation, modulePath)
}
\arguments{
\item{moduleHost}{\code{moduleHost} object}

\item{outputLocation}{\code{outputLocation} object}

\item{modulePath}{output directory on local machine}
}
\description{
Retrieve results of running a module on a remote host
}
\section{Methods (by class)}{
\itemize{
\item \code{vagrantHost}: retrieve module output from
\code{vagrantHost}
}}

