% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runComponent}
\alias{runComponent}
\title{Run a component}
\usage{
runComponent(component, inputs = list(), pipelinePath = getwd())
}
\arguments{
\item{component}{\code{component} object}

\item{inputs}{Named list of absolute paths for component inputs}

\item{pipelinePath}{Pipeline output directory}
}
\value{
FIXME: Result of \code{runModule} or \code{runPipeline}
}
\description{
This function executes a \code{component} with \code{runModule} or
\code{runPipeline} depending on the value of \code{component$value}.
}
\details{
If \code{component$value} is set to \dQuote{module}, the names of
\code{inputs} must match the names in \code{component$value$inputs}.

\code{pipelinePath} must exist on the filesystem.
}
\examples{
mod1 <- module("setX", platform="R",
               sources=list(moduleSource("x <- \\"set\\"")))
pip1 <- pipeline("setX-pipe", description="set the value of x",
                 components=list(mod1))
runComponent(pip1$components$setX)
}

