% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{runComponent}
\alias{runComponent}
\title{Run a component}
\usage{
runComponent(component, inputList = list(), pipelinePath = getwd())
}
\arguments{
\item{component}{\code{component} to be executed}

\item{inputList}{Named list of \code{input} objects}

\item{pipelinePath}{Pipeline output directory}
}
\value{
\code{componentResult} object
}
\description{
This function executes a single pipeline component with \code{runModule} or
\code{runPipeline}.
}
\details{
If the component refers to a \code{module}, the names of
    the elements in \code{inputList} must match the names the
    module's input names.

\code{pipelinePath} must exist on the filesystem.
}

