% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{runComponent}
\alias{runComponent}
\title{Run a component}
\usage{
runComponent(componentName, pipeline, inputObjects = list(),
  pipelinePath = getwd())
}
\arguments{
\item{componentName}{Name of component to be executed}

\item{pipeline}{\code{pipeline} containing component}

\item{inputObjects}{Named list of input objects}

\item{pipelinePath}{Pipeline output directory}
}
\value{
Named list of output objects
}
\description{
This function executes a single pipeline component with \code{runModule} or
\code{runPipeline}.
}
\details{
If the component refers to a \code{module}, the names of
\code{inputs} must match the names the module's input names.

\code{pipelinePath} must exist on the filesystem.
}

