% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runComponent}
\alias{runComponent}
\title{Run a component object}
\usage{
runComponent(component, inputs = list(), pipelinePath = getwd())
}
\arguments{
\item{component}{\code{component} object}

\item{inputs}{Named list of absolute paths for component inputs}

\item{pipelinePath}{Pipeline output directory}
}
\value{
Result of \code{runModule} or \code{runPipeline}
}
\description{
Run a component object
}

