% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runPipeline}
\alias{runPipeline}
\title{Run a pipeline}
\usage{
runPipeline(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} list object}
}
\value{
Meaningless list. TODO: fix what \code{runPipeline},
\code{runModule}, \code{runPlatform} return.
}
\description{
Executes a \code{pipeline}'s \code{component}s.
}
\details{
The function should produce a directory called \file{pipelines} in the
working directory where it will produce its \code{module}s' results.

First the function will determine the order in which its modules are to
be run. Note that the \code{pipeline} is not allowed to have any cycles or
the function will fail.

The function then determines the file paths of the objects described in
the \code{pipe}s as inputs/outputs.

Finally the function runs each \code{module} in the order determined,
feeding each \code{module} its inputs. A directory will be created for each
\code{module} in \file{pipelines/modules}, in which the function will save
the script used to run the \code{module}, and its outputs.
}

