% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{Run a pipeline}
\usage{
runPipeline(pipeline, targetDirectory = getwd())
}
\arguments{
\item{pipeline}{A \code{pipeline} object}

\item{targetDirectory}{File path for pipeline output}
}
\value{
Named list of \code{component} output objects
}
\description{
Executes a \code{pipeline}'s \code{component}s.
}
\details{
This function creates a directory called
\code{pipeline$name} in \file{pipelines}, in the
\code{targetDirectory}. If the directory \file{pipelines} does not
exist in this directory it will be created. Working files and
output from the pipeline's \code{components} will be stored in the
named directory.

First the function will determine the order in which its components are to
be run. Note that the \code{pipeline} is not allowed to have any cycles or
the function will fail.

The function then produces a list of inputs required by the
\code{component}s, and resolves the (intended) location of these.

Finally the function executes each \code{component} in the
determined order.
}
\examples{
simpleGraph <-
    loadPipeline(name = "simpleGraph",
                 ref = system.file("extdata", "simpleGraph",
                                   "simpleGraph-pipeline.xml",
                                    package = "conduit"))

## run the pipeline
runPipeline(simpleGraph, targetDirectory = tempdir())

## python language example
pythonExample <- loadPipeline("pythonExample",
                              system.file("extdata", "pythonExample",
                                          "pipeline.xml", package="conduit"))
runPipeline(pythonExample, targetDirectory = tempdir())

## shell language example
shellExample <- loadPipeline("shellExample",
                             system.file("extdata", "shellExample",
                                         "pipeline.xml", package="conduit"))
runPipeline(shellExample, targetDirectory = tempdir())

## A pipeline with a module run on a remote host
\dontrun{
  irisplots_host_pipeline <- loadPipeline(
      "irisplots_host_pipeline",
      system.file("extdata", "irisplots_host", "pipeline.xml",
      package = "conduit"))
  runPipeline(irisplots_host_pipeline, targetDirectory = tempdir())
}
}
\seealso{
More about \code{pipeline} objects, run single \code{module}
objects with \code{runModule}.
}

