% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runPlatform.shell}
\alias{runPlatform.shell}
\title{Platform support for "shell" platform.}
\usage{
\method{runPlatform}{shell}(module, inputs, modulePath)
}
\arguments{
\item{module}{\code{module} object}

\item{inputs}{Named list of input locations}

\item{modulePath}{File path for module output}
}
\description{
Platform support for "shell" platform.
}
\details{
Creates a .sh script file from the supplied \code{module},
taking specific input file paths from \code{inputs}.

Writes script to \code{modulePath}, then attempts to execute the
script in this location.
}

