% Generated by roxygen2 (4.0.1): do not edit by hand
\name{savePipeline}
\alias{savePipeline}
\title{Save a pipeline to disk}
\usage{
savePipeline(pipeline, targetDirectory = getwd())
}
\description{
Saves a \code{pipeline} to disk as an openapi XML file
}
\details{
As at 2014-08-12 the resulting file is always called \file{pipeline.xml}
}

