% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{savePipeline}
\alias{savePipeline}
\title{Save a pipeline to disk}
\usage{
savePipeline(pipeline, targetDirectory = getwd())
}
\arguments{
\item{pipeline}{\code{pipeline} object}

\item{targetDirectory}{file location to save output}
}
\value{
file path to resulting XML file
}
\description{
Saves a \code{pipeline} to disk as an openapi XML file
}
\details{
As at 2014-08-12 the resulting file is always called \file{pipeline.xml}
}
\examples{
targ1 <- tempdir()
ppl1xml <- system.file("extdata", "simpleGraph", "simpleGraph-pipeline.xml",
		          package = "conduit")
ppl1 <- loadPipeline("simpleGraph",
		        ref = ppl1xml)
savePipeline(pipeline = ppl1, targetDirectory = targ1)
}
\seealso{
\code{pipeline}. For exporting a pipeline and its components see
\code{exportPipeline}
}

