% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{splitPaths}
\alias{splitPaths}
\title{Split a search paths string separated by \code{pathSep}}
\usage{
splitPaths(s)
}
\arguments{
\item{s}{Search paths string}
}
\value{
A list of paths
}
\description{
Split a search paths string separated by \code{pathSep}
}

