% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/pipe.R
\name{start.pipe}
\alias{start.pipe}
\title{Return \code{pipe} start list}
\usage{
\method{start}{pipe}(x)
}
\arguments{
\item{x}{\code{pipe} object}
}
\value{
list containing \code{component} and \code{output}
}
\description{
Return \code{pipe} start list
}

