% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.R
\name{vagrantHost}
\alias{vagrantHost}
\title{Create a \code{vagrantHost} object}
\usage{
vagrantHost(vagrantfile, synchost = dirname(vagrantfile),
  syncguest = "/vagrant")
}
\arguments{
\item{vagrantfile}{Path to vagrantfile}

\item{synchost}{Host synced folder path}

\item{syncguest}{Gues synced folder path}
}
\value{
Object of class \code{vagrantHost} and \code{moduleHost}
}
\description{
Create a \code{moduleHost} object of class \code{vagrantHost} to
describe a vagrant machine for executing module source scripts.
}
\details{
The default setup for vagrant synced folders is for the directory
where the Vagrantfile is found to be synced to \file{/vagrant} on
the vagrant guest. If the vagtant machine describes a different
synced folder setup the synced folder on the host machine can be
specified with \code{synchost}, and on the vagrant guest with
\code{syncguest}.
}

