% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validatePipe}
\alias{validatePipe}
\title{Validate a pipe}
\usage{
validatePipe(p, components)
}
\arguments{
\item{p}{\code{pipe} object}

\item{components}{\code{pipeline} components}
}
\value{
Logical value, TRUE if valid
}
\description{
Validate a pipe
}

