% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validatePipeline}
\alias{validatePipeline}
\title{Validate all pipes in a pipeline}
\usage{
validatePipeline(pipeline)
}
\arguments{
\item{pipeline}{A \code{pipeline} object}
}
\value{
Logical value, TRUE if valid

logical, TRUE if pipeline passes all validity tests
}
\description{
This function checks that all pipes are valid, and returns FALSE if any
are invalid.
}
\details{
A pipe is considered valid if its \code{startComponent} and
\code{endComponent}s values match the value of name \code{component}s in
the pipeline, and if its \code{startOutput} and \code{endInput} values match
a named output or input in the corresponding \code{component}.

A \code{warning} is generated for each invalid \code{pipe} to help debug.
}
\examples{
## load a pipeline
pipel1 <-
    loadPipeline("simpleGraph",
                 system.file("extdata", "simpleGraph",
                             "pipeline.xml",
                             package = "conduit"))
## a valid pipeline
validatePipeline(pipel1)

## break a pipeline
pipel2 <- pipel1
pipel2$pipes[[1]]$start$output <- "veryWrongValue"
## and look, it's invalid!
validatePipeline(pipel2)

}
\seealso{
\code{pipeline}, \code{pipe}
}

