% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.R
\name{vessel}
\alias{fileVessel}
\alias{internalVessel}
\alias{scriptVessel}
\alias{urlVessel}
\alias{vessel}
\title{Functions to create \code{vessel} objects.}
\usage{
internalVessel(symbol)

fileVessel(ref, path = NULL)

urlVessel(ref)

scriptVessel(value)
}
\arguments{
\item{symbol}{symbol name of internal object}

\item{ref}{Absolute or relative location of file for
\code{fileVessel}, URL for \code{urlVessel}}

\item{path}{search path(s) for fileVessel$ref}

\item{value}{Module script as character vector}
}
\value{
\code{vessel} list object
}
\description{
Functions to create vessels for \code{moduleInput},
\code{moduleOutput}, or \code{moduleSource} objects.
}
\details{
\code{vessel} objects are used as sources for
\code{moduleInput} and \code{moduleOutput} objects, or as
destinations for \code{moduleOutput} objects.

\code{internalVessel} objects are used to point to objects
internal to a \code{module} script. They are used to pass objects
internal to a module's platform to other modules.

\code{fileVessel} objects are used to point to files on the local
filesystem.

\code{urlVessel} objects are used to point to objects accessible
via http:// and https:// protocols.

\code{scriptVessel} objects are used to store inline module
source scripts.
}
\examples{
expenses_df <- internalVessel(symbol="expenses")
data_file <- fileVessel(ref="~/myFiles/data.csv")
url_object <- urlVessel(ref = "https://github.com/anhinton/conduit")
source2_script <-
    scriptVessel(
        value=c("print(\\"Hello World!\\")",
                "print(\\"Goodbye cruel world...\\")"))

}
\seealso{
\code{moduleInput}, \code{moduleOutput},
\code{moduleSource}
}

