% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/module.R
\name{vesselToXML}
\alias{vesselToXML}
\title{Create XML corresponding to a \code{vessel} object.}
\usage{
vesselToXML(vessel, namespaceDefinitions = NULL)
}
\arguments{
\item{vessel}{\code{vessel} object}
}
\value{
\code{XMLNode} object
}
\description{
Create XML corresponding to a \code{vessel} object.
}
\seealso{
\code{vessel} objects
}

